% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aih100.R
\docType{data}
\name{aih100}
\alias{aih100}
\title{Banco com 100 registros de um 'arquivo da AIH' (RD??????.DBF).}
\format{

}
\source{
\url{http://www2.datasus.gov.br/DATASUS/index.php?area=0901}
}
\usage{
data(aih100)
}
\description{
Contém todas as variáveis dos "arquivos da AIH" em sua forma atual.
}
\examples{
data("aih100")
str(aih100)

}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2014. SIH – Sistema de Informação Hospitalar do SUS: Manual Técnico Operacional do Sistema. Ministério da Saúde: Brasília, 2015. 87p.
}
\keyword{datasets}
