% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aih500.R
\docType{data}
\name{aih500}
\alias{aih500}
\title{Data frame with selected records from an 'AIH file'}
\format{
An object of class \code{data.frame} with 500 rows and 113 columns.
}
\source{
\url{http://www2.datasus.gov.br/DATASUS/index.php?area=0901}
}
\usage{
data(aih500)
}
\description{
A data frame with 500 records and all variables for the hospital admission form (Autorização de Internação Hospitalar - AIH), the "AIH files" (RD??????.DBC or RD??????.DBF) from DATASUS (Brazil, Ministry of Health), year 2018.
}
\examples{
data("aih500")
str(aih500)
}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2014. SIH – Sistema de Informação Hospitalar do SUS: Manual Técnico Operacional do Sistema. Ministério da Saúde: Brasília, 2015. 87p.
}
\keyword{datasets}
