<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csapAIH.R, R/csapAIHb.R
\name{csapAIH}
\alias{csapAIH}
\title{Classificar Condições Sensíveis à Atenção Primária}
\usage{
csapAIH(
  x,
  grupos = TRUE,
  sihsus = TRUE,
  procobst.rm = TRUE,
  parto.rm = TRUE,
  longa.rm = TRUE,
  cep = TRUE,
  cnes = TRUE,
  arquivo = TRUE,
  sep,
  cid = NULL,
  ...
)

csapAIH(
  x,
  grupos = TRUE,
  sihsus = TRUE,
  procobst.rm = TRUE,
  parto.rm = TRUE,
  longa.rm = TRUE,
  cep = TRUE,
  cnes = TRUE,
  arquivo = TRUE,
  sep,
  cid = NULL,
  ...
)
}
\arguments{
\item{x}{alvo da função: um arquivo, banco de dados ou vetor com códigos da CID-10 (ver \code{detalhes});}

\item{grupos}{argumento lógico, obrigatório; \code{TRUE} (padrão) indica que as internações serão classificadas também em grupos de causas CSAP;}

\item{sihsus}{argumento lógico, obrigatório se \code{x} for um arquivo; \code{TRUE} (padrão) indica que o arquivo ou banco de dados a ser tabulado tem minimamente os seguintes campos dos arquivos da AIH:
\itemize{
  \item{DIAG_PRINC }{diagnóstico principal da internação;}
  \item NASC {data de nascimento; }
  \item DT_INTER {    data da internação; }
  \item DT_SAIDA {    data da alta hospitalar; }
  \item COD_IDADE {    código indicando a faixa etária a que se refere o valor registrado no campo idade; }
  \item IDADE    {idade (tempo de vida acumulado) do paciente, na unidade indicada no campo COD_IDADE;}
  \item MUNIC_RES    {município de residência do paciente; }
  \item MUNIC_MOV {município de internação do paciente; }
  \item SEXO { sexo do paciente; }
  \item N_AIH    { número da AIH; }
  \item {PROC_REA} { procedimento realizado, segundo a tabela do SIH/SUS.}
  }}

\item{procobst.rm}{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) exclui as internações por procedimento obstétrico (\code{ver detalhes});}

\item{parto.rm}{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) exclui as internações por parto (\code{ver detalhes});}

\item{longa.rm}{argumento lógico; \code{TRUE} (padrão) exclui as AIH de longa permanência (AIH tipo 5), retornando uma mensagem com o número e proporção de registros excluídos e o total de registros importados; argumento válido apenas se \code{sihsus=TRUE};}

\item{cep}{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) inclui no banco o Código de Endereçamento Postal do indivíduo; argumento válido apenas se \code{sihsus=TRUE};}

\item{cnes}{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) inclui no banco o nº do hospital no Cadastro Nacional de Estabelecimentos de Saúde; argumento válido apenas se \code{sihsus=TRUE};}

\item{arquivo}{argumento lógico, obrigatório; \code{TRUE} (padrão) indica que o alvo da função (\code{x}) é um arquivo; \code{FALSE} indica que \code{x} é um objeto no espaço de trabalho; é automaticamente marcado como \code{FALSE} quando \code{x} é um \code{factor} ou \code{data frame}; deve ser definido pelo usuário como \code{FALSE} apenas quando \code{x} contiver em seu nome as sequências "dbc", "dbf" ou "csv" sem que isso seja a extensão do arquivo; apenas arquivos com esses formatos podem ser lidos;}

\item{sep}{usado para a leitura de arquivos da AIH em formato CSV; pode ser ";" para arquivos separados por ponto-e-vírgula e com vírgula como separador decimal, ou "," para arquivos separados por vírgula e com ponto como separador decimal;}

\item{cid}{identifica a varivável contendo os códigos da CID-10, em bancos de dados sem a estrutura do SIHSUS; argumento obrigatório nesses casos;}

\item{...}{permite a inclusão de argumentos das funções \code{\link{read.table}} e suas derivadas.}
}
\value{
A função tem diferentes possibilidades de retorno, segundo a estrutura dos dados lidos e as opções de leitura:
\itemize{
 \item{ Se for um arquivo ou \code{data frame} com a estrutura dos arquivos da AIH:
   \itemize{
     \item um \code{data frame} com as variáveis nº da AIH, município de residência, município de internação, sexo, data de nascimento, idade em anos completos, faixa etária detalhada, faixa etária quinquenal, data da internação, data da saída, procedimento realizado, cid, CSAP, grupo csap, CEP e CNES do hospital
       \itemize{
         \item Nesse caso, o banco resultante tem um argumento "resumo" com o resumo da importação de dados, segundo as opções de seleção
         \item{ Se os argumentos \code{grupo}, \code{cep} ou \code{cnes} forem definidos como \code{FALSE}, o banco é construído sem essas variáveis}
     }}}
   \item{Se um fator ou data frame sem a estrutura dos arquivos da AIH:}
 \itemize{
 \item{Se \code{grupos = TRUE}: }{ um banco de dados com as variáveis \code{csap} (sim ou não), \code{grupo} (subgrupo CSAP) e \code{cid} (código da CID-10)};
 \item{Se \code{factor} e \code{grupos = FALSE}: }{ um fator com as observações classificadas como CSAP ou não-CSAP}.    }
 }

A função tem diferentes possibilidades de retorno, segundo a estrutura dos dados lidos e as opções de leitura:
\itemize{
 \item{ Se for um arquivo ou \code{data frame} com a estrutura dos arquivos da AIH:
   \itemize{
     \item um \code{data frame} com as variáveis nº da AIH, município de residência, município de internação, sexo, data de nascimento, idade em anos completos, faixa etária detalhada, faixa etária quinquenal, data da internação, data da saída, procedimento realizado, cid, CSAP, grupo csap, CEP e CNES do hospital
       \itemize{
         \item Nesse caso, o banco resultante tem um argumento "resumo" com o resumo da importação de dados, segundo as opções de seleção
         \item{ Se os argumentos \code{grupo}, \code{cep} ou \code{cnes} forem definidos como \code{FALSE}, o banco é construído sem essas variáveis}
     }}}
   \item{Se um fator ou data frame sem a estrutura dos arquivos da AIH:}
 \itemize{
 \item{Se \code{grupos = TRUE}: }{ um banco de dados com as variáveis \code{csap} (sim ou não), \code{grupo} (subgrupo CSAP) e \code{cid} (código da CID-10)};
 \item{Se \code{factor} e \code{grupos = FALSE}: }{ um fator com as observações classificadas como CSAP ou não-CSAP}.    }
 }
}
\description{
Classifica códigos da 10ª Revisão da Classificação Internacional de Doenças (CID-10) segundo a Lista Brasileira de Internação por Condições Sensíveis à Atenção Primária e oferece outras funcionalidades, especialmente para o manejo dos "arquivos da AIH" (RD??????.DBC; BD-SIH/SUS).

Classifica códigos da 10ª Revisão da Classificação Internacional de Doenças (CID-10) segundo a Lista Brasileira de Internação por Condições Sensíveis à Atenção Primária e oferece outras funcionalidades, especialmente para o manejo dos "arquivos da AIH" (RD??????.DBC; BD-SIH/SUS).
}
\details{
\itemize{
  \item{x pode ser:}{
   \enumerate{
    \item{ }{um arquivo de dados armazenado num diretório; }
    \item{ }{um banco de dados, ou um vetor da classe \code{factor} presente como objeto no espaço de trabalho do R, em que uma das variáveis, ou o vetor, contenha códigos da CID-10. }
   }

Se for um arquivo, o nome deve ser escrito entre aspas e com a extensão do arquivo (DBC, DBF ou CSV, em minúsculas ou maíusculas). Se não estiver no diretório de trabalho ativo, seu nome deve ser precedido pelo caminho (path) até o diretório de armazenamento. Se estiver em outro formato, podem-se usar os argumentos da função \code{\link{read.table}} para leitura dos dados.

Se a função for dirigida a um objeto no espaço de trabalho da classe \code{factor} ou \code{data.frame}, estes também são reconhecidos e o comando é o mesmo: \code{csapAIH(<objeto>)}. Se o objeto for de outra classe, como \code{character} ou \code{matrix}, é necessário definir o argumento "arquivo" como FALSE: \code{csapAIH(<objeto>, arquivo = FALSE)}, ou, para vetores isolados, defini-lo como fator: \code{csapAIH(as.factor(<objeto>))}.
}

\item{procbst.rm } {= TRUE (padrão) exclui as internações por procedimentos relacionados ao parto ou abortamento. São excluídas as internações pelos seguintes procedimentos obstétricos, independente do diagnóstico:
  \itemize{
   \item{0310010012 } ASSISTENCIA AO PARTO S/ DISTOCIA
   \item{0310010020 } ATENDIMENTO AO RECÉM-NASCIDO EM SALA DE PARTO
   \item{0310010039 } PARTO NORMAL
   \item{0310010047 } PARTO NORMAL EM GESTAÇÃO DE ALTO RISCO
   \item{0411010018 } DESCOLAMENTO MANUAL DE PLACENTA
   \item{0411010026 } PARTO CESARIANO EM GESTAÇÃO ALTO RISCO
   \item{0411010034 } PARTO CESARIANO
   \item{0411010042 } PARTO CESARIANO C/ LAQUEADURA TUBÁRIA
   \item{0411020013 } CURETAGEM PÓS-ABORTAMENTO / PUERPERAL
   \item{0411020021 } EMBRIOTOMIA
   } }
\item{parto.rm } {= TRUE (padrão) exclui as internações por parto pelo campo diagnóstico, independente do procedimento. São excluídas as internações com os seguintes diagnósticos (CID-10):
 \itemize{
  \item{O80} Parto único espontâneo
  \item{O81} Parto único por fórceps ou vácuo-extrator
  \item{O82} Parto único por cesariana
  \item{O83} Outros tipos de parto único assistido
  \item{O84} Parto múltiplo
 }
}

É retornada uma mensagem informando o número de registros lidos, o número e proporção de registros excluídos e o total de registros importados.

\item{\code{sihsus}  }{ A própria função define este argumento como \code{FALSE} quando "x" (o alvo da função) é um fator. Quando o alvo é um objeto da clase \code{data frame} sem a estrutura dos arquivos da AIH, a variável com os códigos da CID-10 deve ser trabalhada como um \code{factor}. }
}

\itemize{
  \item{x pode ser:}{
   \enumerate{
    \item{ }{um arquivo de dados armazenado num diretório; }
    \item{ }{um banco de dados, ou um vetor da classe \code{factor} presente como objeto no espaço de trabalho do R, em que uma das variáveis, ou o vetor, contenha códigos da CID-10. }
   }

Se for um arquivo, o nome deve ser escrito entre aspas e com a extensão do arquivo (DBC, DBF ou CSV, em minúsculas ou maíusculas). Se não estiver no diretório de trabalho ativo, seu nome deve ser precedido pelo caminho (path) até o diretório de armazenamento. Se estiver em outro formato, podem-se usar os argumentos da função \code{\link{read.table}} para leitura dos dados.

Se a função for dirigida a um objeto no espaço de trabalho da classe \code{factor} ou \code{data.frame}, estes também são reconhecidos e o comando é o mesmo: \code{csapAIH(<objeto>)}. Se o objeto for de outra classe, como \code{character} ou \code{matrix}, é necessário definir o argumento "arquivo" como FALSE: \code{csapAIH(<objeto>, arquivo = FALSE)}, ou, para vetores isolados, defini-lo como fator: \code{csapAIH(as.factor(<objeto>))}.
}

\item{procbst.rm } {= TRUE (padrão) exclui as internações por procedimentos relacionados ao parto ou abortamento. São excluídas as internações pelos seguintes procedimentos obstétricos, independente do diagnóstico:
  \itemize{
   \item{0310010012 } ASSISTENCIA AO PARTO S/ DISTOCIA
   \item{0310010020 } ATENDIMENTO AO RECÉM-NASCIDO EM SALA DE PARTO
   \item{0310010039 } PARTO NORMAL
   \item{0310010047 } PARTO NORMAL EM GESTAÇÃO DE ALTO RISCO
   \item{0411010018 } DESCOLAMENTO MANUAL DE PLACENTA
   \item{0411010026 } PARTO CESARIANO EM GESTAÇÃO ALTO RISCO
   \item{0411010034 } PARTO CESARIANO
   \item{0411010042 } PARTO CESARIANO C/ LAQUEADURA TUBÁRIA
   \item{0411020013 } CURETAGEM PÓS-ABORTAMENTO / PUERPERAL
   \item{0411020021 } EMBRIOTOMIA
   } }
\item{parto.rm } {= TRUE (padrão) exclui as internações por parto pelo campo diagnóstico, independente do procedimento. São excluídas as internações com os seguintes diagnósticos (CID-10):
 \itemize{
  \item{O80} Parto único espontâneo
  \item{O81} Parto único por fórceps ou vácuo-extrator
  \item{O82} Parto único por cesariana
  \item{O83} Outros tipos de parto único assistido
  \item{O84} Parto múltiplo
 }
}

É retornada uma mensagem informando o número de registros lidos, o número e proporção de registros excluídos e o total de registros importados.

\item{\code{sihsus}  }{ A própria função define este argumento como \code{FALSE} quando "x" (o alvo da função) é um fator. Quando o alvo é um objeto da clase \code{data frame} sem a estrutura dos arquivos da AIH, a variável com os códigos da CID-10 deve ser trabalhada como um \code{factor}. }
}
}
\note{
A função \code{\link{read.dbf}}, do pacote \code{foreign}, não lê arquivos em formato DBF em que uma das variáveis tenha todos os valores ausentes ('missings'); essas variáveis devem ser excluídas antes da leitura do arquivo pela função \code{csapAIH} ou mesmo pelas função \code{\link{read.dbf}}.

A função \code{\link{read.dbf}}, do pacote \code{foreign}, não lê arquivos em formato DBF em que uma das variáveis tenha todos os valores ausentes ('missings'); essas variáveis devem ser excluídas antes da leitura do arquivo pela função \code{csapAIH} ou mesmo pelas função \code{\link{read.dbf}}.
}
\examples{
## Uma lista de códigos da CID-10:
##---------------------------------
cids <- c("I200", "K929", "T16", "I509", "I10",  "I509", "S068")
teste1 <- csapAIH(as.factor(cids)) ; class(teste1) ; teste1
teste2 <- csapAIH(as.factor(cids),  grupo=FALSE) ; class(teste2) ; teste2

## Um 'arquivo da AIH' armazenado no diretório de trabalho:
##---------------------------------------------------------
\dontrun{
 teste3.dbf <- csapAIH("RDRS1301.dbf")
 str(teste3.dbf)
 teste3.dbc <- csapAIH("RDRS1301.dbc")
 str(teste3.dbc)
}

## Um 'data.frame' com a estrutura dos 'arquivos da AIH':
##-------------------------------------------------------
data("aih100")
str(aih100)
teste4 <- csapAIH(aih100)
str(teste4)

## Uma base de dados com a estrutura dos 'arquivos da AIH'
## mas sem as variáveis CEP ou CNES:
##--------------------------------------------------------
aih <- subset(aih100, select = -c(CEP, CNES))
teste5 <- csapAIH(aih, cep = FALSE, cnes = FALSE)
str(teste5)

## Para uma base de dados sem a estrutura dos BD-SIH/SUS, apenas trabalhe
## a variável com os CIDs, como nos primeiros exemplos (teste1 e teste2)
##-----------------------------------------------------------------------
## teste6 <- csapAIH(BaseDeDados$VariavelcomCID)

## Uma lista de códigos da CID-10:
##---------------------------------
cids <- c("I200", "K929", "T16", "I509", "I10",  "I509", "S068")
teste1 <- csapAIH(as.factor(cids)) ; class(teste1) ; teste1
teste2 <- csapAIH(as.factor(cids),  grupo=FALSE) ; class(teste2) ; teste2

## Um 'arquivo da AIH' armazenado no diretório de trabalho:
##---------------------------------------------------------
\dontrun{
 teste3.dbf <- csapAIH("RDRS1301.dbf")
 str(teste3.dbf)
 teste3.dbc <- csapAIH("RDRS1301.dbc")
 str(teste3.dbc)
}

## Um 'data.frame' com a estrutura dos 'arquivos da AIH':
##-------------------------------------------------------
data("aih100")
str(aih100)
teste4 <- csapAIH(aih100)
str(teste4)

## Uma base de dados com a estrutura dos 'arquivos da AIH'
## mas sem as variáveis CEP ou CNES:
##--------------------------------------------------------
aih <- subset(aih100, select = -c(CEP, CNES))
teste5 <- csapAIH(aih, cep = FALSE, cnes = FALSE)
str(teste5)

## Para uma base de dados sem a estrutura dos BD-SIH/SUS, apenas trabalhe
## a variável com os CIDs, como nos primeiros exemplos (teste1 e teste2)
##-----------------------------------------------------------------------
## teste6 <- csapAIH(BaseDeDados$VariavelcomCID)

}
\references{
Alfradique et al., Internações por Condições Sensíveis à Atenção Primária: a construção da lista brasileira como ferramenta para medir o desempenho do sistema de saúde (Projeto ICSAP - Brasil). Cad Saúde Pública 25(6):1337-49.

Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}

__________. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2010. Manual técnico operacional do Sistema de Informação Hospitalar: orientações técnicas. Versão 01.2013. Ministério da Saúde: Brasília, 2013.

Alfradique et al., Internações por Condições Sensíveis à Atenção Primária: a construção da lista brasileira como ferramenta para medir o desempenho do sistema de saúde (Projeto ICSAP - Brasil). Cad Saúde Pública 25(6):1337-49.

Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}

__________. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2010. Manual técnico operacional do Sistema de Informação Hospitalar: orientações técnicas. Versão 01.2013. Ministério da Saúde: Brasília, 2013.
}
\seealso{
\code{\link{read.table}}, \code{\link{read.csv}}, \code{\link[read.dbc]{read.dbc}},
\code{\link{descreveCSAP}}, \code{\link{desenhaCSAP}}, \code{\link{nomesgruposCSAP}}

\code{\link{read.table}}, \code{\link{read.csv}}, \code{\link[read.dbc]{read.dbc}},
\code{\link{descreveCSAP}}, \code{\link{desenhaCSAP}}, \code{\link{nomesgruposCSAP}}
}
\keyword{AIH-SUS;}
\keyword{CSAP;}
\keyword{package}
=======
\name{csapAIH}
\alias{csapAIH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
% \packageTitle{csapAIH}
Classificar Condicoes Sensiveis a Atencao Primaria
}
\description{
% \section{Descrição}{ 
Classifica códigos da 10ª Revisão da Classificação Internacional de Doenças (CID-10) segundo a Lista Brasileira de Internação por Condições Sensíveis à Atenção Primária e oferece outras funcionalidades, especialmente para o manejo dos "arquivos da AIH" (RD??????.DBF; BD-SIH/SUS). 
% }
}

\section{Uso}{
csapAIH(x, grupos = TRUE, 
        sihsus = TRUE, x.procobst = TRUE, longa = FALSE,
        cep = TRUE, cnes = TRUE, arquivo = TRUE, sep, ...)
}

\section{Argumentos}{
\itemize{
  \item{\code{x}  }{alvo da função: um arquivo, banco de dados ou vetor com códigos da CID-10 (ver \code{detalhes}); }
  \item{\code{grupos}  }{argumento lógico, obrigatório; \code{TRUE} (padrão) indica que as internações serão classificadas também em grupos de causas CSAP; }
  \item{\code{sihsus}  }{argumento lógico, obrigatório se \code{x} for um arquivo; \code{TRUE} (padrão) indica que o arquivo ou banco de dados a ser tabulado tem minimamente os seguintes campos dos arquivos da AIH:
\itemize{
\item{DIAG_PRINC }{diagnóstico principal da internação;}
\item NASC {data de nascimento; }
\item DT_INTER {	data da internação; }
\item DT_SAIDA {	data da alta hospitalar; }
\item COD_IDADE {	código indicando a faixa etária a que se refere o valor registrado no campo idade; }
\item IDADE	{idade (tempo de vida acumulado) do paciente, na unidade indicada no campo COD_IDADE;}
\item MUNIC_RES	{município de residência do paciente; }
\item MUNIC_MOV {município de internação do paciente; }
\item SEXO { sexo do paciente; }
\item N_AIH	{ número da AIH; }
\item { PROC_REA	procedimento realizado, segundo a tabela do SIH/SUS; }
} }
  \item{\code{x.procobst}:  }{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) exclui as internações por procedimentos relacionados ao parto ou abortamento (\code{ver detalhes}); }
  \item{\code{longa}: }{argumento lógico; \code{FALSE} (padrão) exclui as AIH de longa permanência (AIH tipo 5), retornando uma mensagem com o número e proporção de registros excluídos e o total de registros importados; argumento válido apenas se \code{sihsus=TRUE}; }
  \item{\code{cep}:  }{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) inclui no banco o Código de Endereçamento Postal do indivíduo; argumento válido apenas se \code{sihsus=TRUE}; }
  \item{\code{cnes}:  }{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) inclui no banco o nº do hospital no Cadastro Nacional de Estabelecimentos de Saúde; argumento válido apenas se \code{sihsus=TRUE}; }
  \item{\code{arquivo}:  }{argumento lógico, obrigatório; \code{TRUE} (padrão) indica que o alvo da função (\code{x}) é um arquivo; \code{FALSE} indica que \code{x} é um objeto no espaço de trabalho; é automaticamente marcado como \code{FALSE} quando \code{x} é um \code{factor} ou \code{data frame}; deve ser definido pelo usuário como \code{FALSE} apenas quando \code{x} contiver sem seu nome as sequências "dbf" ou "csv" sem que isso seja a extensão do arquivo; apenas arquivos com a estrutura citada acima podem ser lidos; }
  \item{\code{sep}: }{ usado para a leitura de arquivos da AIH em formato CSV; pode ser ";" para arquivos separados por ponto-e-vírgula e com vírgula como separador decimal, ou "," para arquivos separados por vírgula e com ponto como separador decimal. }
}
}

\section{Detalhes}{
\itemize{  
  \item{x pode ser:}{ 
  \enumerate{ \item{ }{um arquivo de dados armazenado num diretório; } 
              \item{ }{um banco de dados, ou um vetor da classe \code{factor} presente como objeto no espaço de trabalho do R, em que uma das variáveis, ou o vetor, contenha códigos da CID-10. }
              }

  Se for um arquivo, o nome deve ser escrito entre aspas e com a extensão do arquivo (DBF ou CSV, em minúsculas ou maíusculas). Se não estiver no diretório de trabalho ativo, seu nome deve ser precedido pelo caminho (path) até o diretório de armazenamento. Se estiver em outro formato, podem-se usar os argumentos da função \code{\link{read.table}} para leitura dos dados.
  
  Se a função for dirigida a um objeto no espaço de trabalho da classe \code{factor} ou \code{data.frame}, estes também são reconhecidos e o comando é o mesmo: \code{csapAIH(<objeto>)}. Se o objeto for de outra classe, como \code{character} ou \code{matrix}, é necessário definir o argumento "arquivo" como FALSE: \code{csapAIH(<objeto>, arquivo = FALSE)}, ou, para vetores isolados, defini-lo como fator: \code{csapAIH(as.factor(<objeto>))}.
  
  É retornada uma mensagem informando o número de registros lidos.
  }
\item{x.procbst } {= TRUE (padrão) exclui as internações pelos seguintes procedimentos obstétricos:
\itemize{
\item{0310010012 } ASSISTENCIA AO PARTO S/ DISTOCIA
\item{0310010020 } ATENDIMENTO AO RECÉM-NASCIDO EM SALA DE PARTO
\item{0310010039 } PARTO NORMAL
\item{0310010047 } PARTO NORMAL EM GESTAÇÃO DE ALTO RISCO 
\item{0411010018 } DESCOLAMENTO MANUAL DE PLACENTA
\item{0411010026 } PARTO CESARIANO EM GESTAÇÃO ALTO RISCO
\item{0411010034 } PARTO CESARIANO
\item{0411010042 } PARTO CESARIANO C/ LAQUEADURA TUBÁRIA
\item{0411020013 } CURETAGEM PÓS-ABORTAMENTO / PUERPERAL
\item{0411020021 } EMBRIOTOMIA
}
 É retornada uma mensagem informando o número e proporção de registros excluídos e o total de registros importados.
 }
  \item{\code{sihsus}  }{ A própria função define este argumento como \code{FALSE} quando "x" (o alvo da função) é um fator. Quando o alvo é um objeto da clase \code{data frame} sem a estrutura dos arquivos da AIH, a variável com os códigos da CID-10 deve ser trabalhada como um \code{factor}. }
} }

\section{Valor}{
A função tem diferentes possibilidades de retorno, segundo a estrutura dos dados lidos e as opções de leitura:
\itemize{
  \item{ Se for um arquivo ou \code{data frame} com a estrutura dos arquivos da AIH: um \code{data frame} com as variáveis nº da AIH, município de residência, município de internação, sexo, data de nascimento, idade em anos completos, faixa etária detalhada, faixa etária quinquenal, data da internação, data da saída, procedimento realizado, cid, CSAP, grupo csap, CEP e CNES do hospital}
  \itemize{\item{Se os argumentos \code{grupo}, \code{cep} ou \code{cnes} forem definidos como \code{FALSE}, o banco é construído sem essas variáveis}}
  \item{Se um fator ou data frame sem a estrutura dos arquivos da AIH:}
   \itemize{
    \item{Se \code{grupos = TRUE}: }{ um banco de dados com as variáveis \code{csap} (sim ou não), \code{grupo} (subgrupo CSAP) e \code{cid} (código da CID-10)};
        \item{Se \code{factor} e \code{grupos = FALSE}: }{ um fator com as observações classificadas como CSAP ou não-CSAP}.    }
  } 
} 

\section{Autor}{
Fúlvio B. Nedel \email{fulvio.nedel@ufsc.br}
\itemize{\item Departamento de Saúde Pública, CCS/UFSC
          \url{http://spb.ufsc.br/}
         \item Grups de Recerca d'Amèrica i Àfrica Llatines, GRAAL-UAB
          \url{http://graal.uab.cat}
         }
}

\section{Referências}{
Alfradique et al., Internações por Condições Sensíveis à Atenção Primária: a construção da lista brasileira como ferramenta para medir o desempenho do sistema de saúde (Projeto ICSAP - Brasil). Cad Saúde Pública 25(6):1337-49.

Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}

__________. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2010. Manual técnico operacional do Sistema de Informação Hospitalar: orientações técnicas. Versão 01.2013. Ministério da Saúde: Brasília, 2013.
}

\note{
A função \code{\link{read.dbf}}, do pacote \code{foreign}, não lê arquivos em formato DBF em que uma das variáveis tenha todos os valores ausentes ('missings'); essas variáveis devem ser excluídas antes da leitura do arquivo pela função \code{csapAIH} ou mesmo pelas função \code{\link{read.dbf}}.
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
\code{\link{read.table}}  \code{\link{read.csv}}
}

\examples{
## Uma lista de códigos da CID-10:
##---------------------------------
cids <- c("I200", "K929", "T16", "I509", "I10",  "I509", "S068")
teste1 <- csapAIH(as.factor(cids)) ; class(teste1) ; teste1
teste2 <- csapAIH(as.factor(cids),  grupo=FALSE) ; class(teste2) ; teste2

## Um 'arquivo da AIH' armazenado no diretório de trabalho:
##---------------------------------------------------------
\dontrun{
teste3 <- csapAIH("RDRS1301.dbf")
str(teste3)
}

## Um 'data.frame' com a estrutura dos 'arquivos da AIH':
##-------------------------------------------------------
data("aih100")
str(aih100)
teste4 <- csapAIH(aih100)
str(teste4)

## Uma base de dados com a estrutura dos 'arquivos da AIH' 
## mas sem as variáveis CEP ou CNES:
##--------------------------------------------------------
aih <- subset(aih100, select = -c(CEP, CNES))
teste5 <- csapAIH(aih, cep = FALSE, cnes = FALSE)
str(teste5)

## Para uma base de dados sem a estrutura dos BD-SIH/SUS, apenas 
## trabalhe a variável com os CIDs, como nos primeiros exemplos
##--------------------------------------------------------------
## teste6 <- csapAIH(BaseDeDados$VariavelcomCID)

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ CSAP }
\keyword{ AIH-SUS }
\keyword{ package }
\keyword{ function }% __ONLY ONE__ keyword per line
>>>>>>> 29c50c2708d01f4a9cb0381092092311f11ae1c9
