% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descreveCSAP.R
\name{descreveCSAP}
\alias{descreveCSAP}
\title{Tabular Condições Sensíveis à Atenção Primária}
\usage{
descreveCSAP(grupos, digits = 2, lang = "pt.ca")
}
\arguments{
\item{grupos}{Pode ser:
\itemize{
  \item Um \code{data.frame} gerado pela função \code{\link{csapAIH}}, ou qualquer \code{data.frame} com uma variável chamada \code{grupo} com os grupos de causa da Lista Brasileira de Internações por CSAP, rotulados como os resultantes da função \code{\link{csapAIH}}, isto é, "g01", "g02", ..., "g19".
  \item Um vetor da classe \code{factor} ou \code{character} com os grupos de causa CSAP, nomeados de acordo com o resultado da função \code{\link{csapAIH}}. Esse vetor não precisa ser gerado pela função \code{\link{csapAIH}}, mas deve conter todos os 19 grupos de causa, ainda que sua frequência seja zero, e também devem ser rotulados da mesma forma e ordem que na função, isto é, "g01", "g02", ..., "g19".
  }}

\item{digits}{Número de decimais nas proporções apresentadas.}

\item{lang}{Define o idioma dos nomes dos grupos. O padrão é \code{"pt.ca"} ("português com acentos"). Pode ser \code{"pt.sa"}, \code{"es"} ou \code{"en"}.}
}
\value{
Um objeto da classe \code{data.frame} com a tabulação dos códigos da CID-10 segundo os grupos de causa da Lista Brasileira de ICSAP (19 grupos), com a frequência absoluta de casos e as porcentagens sobre o total de internações e sobre o conjunto das ICSAP.
}
\description{
Constrói uma tabela de frequências absolutas e relativas das CSAP por grupo de causa
}
\note{
Os valores são armazenados em formato latino (vírgula como separador decimal e ponto como separador de milhar) e são, portanto, da classe \code{character}. Isso é indesejado, porque impede a realização de operações matemáticas ou mesmo a aplicação direta da função \code{\link{as.numeric}}, para retornar à classe \code{numeric}. Para resolver esse problema foi criada uma nova função \code{\link{tabCSAP}} que tem por output uma versão da tabela com os valores numéricos, com um argumento para a apresentação em formato latino. Portanto, para evitar a quebra de código já escrito, \code{\link{descreveCSAP}} será mantida como está, mas não será mais desenvolvida com esse nome, sua continuidade se dará por meio de \code{\link{tabCSAP}}. Tabulações com a lista de 20 grupos de causa deve ser feita com esta última função.
}
\examples{
data(aih100)
df = csapAIH(aih100)
descreveCSAP(df$grupo)
descreveCSAP(df)

df = csapAIH(eeh20, sihsus = FALSE, cid = cau_cie10, lista = "Alfradique")
descreveCSAP(df$grupo)
descreveCSAP(df, lang = "es")

}
\seealso{
\code{\link{csapAIH}}, \code{\link{desenhaCSAP}}, \code{\link{nomesgruposCSAP}}, \code{\link{tabCSAP}}
}
