% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desenhaCSAP.R
\name{desenhaCSAP}
\alias{desenhaCSAP}
\title{Grafico das Condicoes Sensiveis a Atencao Primaria}
\usage{
desenhaCSAP(
  dados,
  lista = "MS",
  lang = "pt.ca",
  jaetabela = FALSE,
  tipo.graf = "ggplot",
  valores = "porcento",
  ordenar = TRUE,
  colorir = TRUE,
  porcentagens = TRUE,
  val.dig = 0,
  titulo = NULL,
  onde,
  quando = NULL,
  t.hjust = 1,
  t.size = 12,
  x.size = 10,
  y.size = 11,
  val.size = 2.5,
  limsup = NULL,
  ...
)
}
\arguments{
\item{dados}{O objeto com as informações a serem desenhadas. Pode ser: (ver \code{\link{descreveCSAP}})
\itemize{
  \item Um \code{data.frame} gerado pela função \code{\link{csapAIH}}, ou qualquer \code{data.frame} com uma variável chamada \code{grupo} com os grupos de causa da Lista Brasileira de CSAP, rotulados na mesma forma que os resultantes da função \code{\link{csapAIH}}, isto é, "g01", "g02", ..., "g19".
  \item Um objeto da classe \code{factor}) ou \code{character} com os grupos de causa CSAP, em ordem crescente de 1 a 19, conforme os grupos da Portaria do MS.
  nomeados de acordo com o resultado da função \code{\link{csapAIH}}. Esse vetor não precisa ser gerado pela função \code{\link{csapAIH}}, mas os grupos também devem ser rotulados da mesma forma que na função, isto é, "g01", "g02", ..., "g19".
  }}

\item{lista}{Lista de causas a ser considerada (v. detalhes); pode ser \code{"MS"} (padrão) para a lista publicada em portaria pelo Ministério da Saúde do Brasil ou "Alfradique" para a lista publicada no artigo de Alfradique et al.}

\item{lang}{idioma em que se apresentam os nomes dos grupos; pode ser: "pt.ca" (default) para nomes em português com acentos; "pt.sa" para nomes em português sem acentos; "en" para nomes em inglês; ou "es" para nomes em castelhano.}

\item{jaetabela}{Argumento lógico, cujo padrão é FALSE. TRUE indica que os dados são uma tabela pronta, que deve apenas ser graficada. A tabela pode ser um objeto de qualquer classe representando uma tabela com pelo menos duas colunas, sendo a primeira com uma identificação (não necessariamente o nome) do grupo CSAP conforme a Lista Brasileira e a segunda com o número de casos observado em cada grupo.}

\item{tipo.graf}{"ggplot" (padrão) cria um gráfico com \code{\link[ggplot2]{ggplot2}}; quando definido como "base", ou quando \code{\link[ggplot2]{ggplot2}} não está instalado, desenha um gráfico com as funções básicas.}

\item{valores}{Argumento utilizado nos gráficos com \code{\link[ggplot2]{ggplot2}}; "porcento" (padrão) desenha as barras em porcentagem, "contagem" as desenha em frequência absoluta. Veja em 'detalhes'.}

\item{ordenar}{Argumento lógico. Se \code{TRUE} (padrão), as barras do gráfico serão ordenadas de maior a menor frequência de casos.}

\item{colorir}{Argumento lógico ou para colorir as barras segundo demanda.
\itemize{
 \item \code{TRUE} (padrão), colore as barras com a paleta \code{\link{rainbow}};
 \item \code{FALSE}, mantém como \code{NULL} os argumentos para a cor das barras nas funções de origem do gráfico, isto é \code{\link[graphics]{barplot}(..., col = NULL, ...)} e \cr \code{\link[ggplot2]{geom_bar}(fill = NULL)}, colorindo as barras de \code{barplot} em cinza e de \code{ggplot} em preto;
 \item \code{"cinza"} colore as barras em quatro tons de cinza de escuro a claro: os cinco grupos mais frequentes, os cinco seguintes, ..., os quatro últimos
 \item Aceita um vetor (de comprimento igual ao nº de grupos de causa CSAP (o nº de \code{levels} na variável \code{grupos})) com os nomes ou códigos de cores.
 }}

\item{porcentagens}{Argumento lógico, válido apenas para gráficos com ggplot. Se \code{TRUE} (padrão), as barras terão escritas sobre elas a porcentagem do grupo de causa sobre o total de internações (ou o total de internações no estrato, em gráficos com \code{\link[ggplot2]{facet_grid}} ou \code{\link[ggplot2]{facet_wrap}}).}

\item{val.dig}{Nº de decimais nos valores das barras; padrão é 0.}

\item{titulo}{Título do gráfico; se NULL (default), não é gerado um título; se \code{"auto"}, o argumento \code{onde} passa a ser obrigatório e a função gera um título para o gráfico a partir da informação de \code{onde} e do arquivo de dados ou do informado para o argumento \code{quando}. Se o argumento \code{dados} for um \code{factor} ou \code{character}, o argumento \code{quando} é obrigatório.}

\item{onde}{Local, população de origem dos dados do gráfico; obrigatório se \code{titulo = "auto"}.}

\item{quando}{Período de referência dos dados; se a fonte de dados for um "arquivo da AIH" (RD??????.dbc), é automaticamente extraído do arquivo.}

\item{t.hjust}{Valor para definição de ajuste horizontal do título, válido apenas para gráficos com ggplot. Default é 1.}

\item{t.size}{Valor para definição do tamanho de letra do título, válido apenas para gráficos com ggplot. Default é 12.}

\item{x.size}{Tamanho da letra do eixo x, válido apenas para gráficos com ggplot. Default é 10.}

\item{y.size}{Tamanho da letra do eixo y, válido apenas para gráficos com ggplot. Default é 12.}

\item{val.size}{Tamanho da letra dos valores das barras. Padrão é 2.5.}

\item{limsup}{Valor para ajuste do espaçamento do eixo de frequências, válido apenas para gráficos com ggplot. Quando o eixo representa porcentagens, deve ser expresso em proporção.}

\item{...}{Permite o uso de argumentos de \code{\link{plot}} e \code{\link{barplot}}}
}
\value{
Na opção padrão e com \code{\link[ggplot2]{ggplot2}} instalado, devolve um objeto das classes "gg" e "ggplot", com o gráfico.
}
\description{
Desenha um gráfico de barras das CSAP por grupo de causa segundo a Lista Brasileira de Internações por Condições Sensíveis à Atenção Primária. Permite a lista oficial publicada em Portaria Ministerial, com 19 grupos de causa, ou a lista com 20 grupos, publicada por Alfradique et al.
}
\details{
O gráfico é desenhado com \code{\link[ggplot2]{ggplot2}}. Portanto, segue essa filosofia e permite a adição de outros comandos ao objeto devolvido. O vetor \code{grupos} não precisa ser gerado com a função \code{\link{csapAIH}}, mas deve usar os mesmos caracteres de identificação dos grupos CSAP que o resultado da função, v.g. "g01", "g02", ..., "g19".
}
\examples{
library(csapAIH)
data("aih100") # Carregar o banco de dados de exemplo
df   <- csapAIH(aih100) # Computar as CSAP, lista MS

#  Graficos com ggplot
# =====================

# Cria o grafico a partir do banco,
# uma vez que a variavel com os grupos se chama "grupo":
# ----------------------------------------------------------
# Com titulo "automatico":
desenhaCSAP(df, titulo = "auto", onde = "Rio Grande do Sul")

# Sem titulo e sem ordenacacao por frequencia:
desenhaCSAP(df, ordenar = FALSE)
#
# Com a lista de Alfradique et al.:
desenhaCSAP(csapAIH(aih100, "Alfradique"),
            lista = "Alfradique",
            titulo = "auto",
            onde = "Rio Grande do Sul")
desenhaCSAP(csapAIH(aih100, "Alfradique"), lista = "Alfradique", lang = "es")

# Cores
#-------
# Sem cores nas barras
desenhaCSAP(df, colorir = FALSE)

# Com as barras em tons de cinza
desenhaCSAP(df, colorir = "cinza")

# Com as barras em outra cor
desenhaCSAP(df, colorir = "yellow")
#
# Usando o banco todo pode-se tirar proveito de facilidades do ggplot2,
# como a reproducao do grafico por estratos de outras variaveis, como no
# exemplo abaixo com o sexo. Para isso temos de descolorir o grafico, que
# pode ser novamente colorido mais tarde. Como a ordenacao dos grupos de
# causa continua sendo feita pela frequencia da distribuicao global, aqui
# ela faz menos sentido.
  desenhaCSAP(df, ordenar = FALSE) +
    ggplot2::facet_wrap(~sexo)

# Cria o grafico a partir de uma variavel:
# ---------------------------------------
fator <- df$grupo
desenhaCSAP(fator)
carater <- as.character(fator)
desenhaCSAP(carater, limsup = 4.4)

# Se \code{titulo = "auto"}, o argumento \code{quando} eh obrigatorio:
\dontrun{
 desenhaCSAP(carater, titulo = "auto", onde = 'RS') # resulta em erro
 }
 desenhaCSAP(carater, titulo = "auto", onde = "RS", quando = "jan/2012")
 desenhaCSAP(carater, titulo = "Título manual")

# Cria o grafico a partir de uma tabela com a primeira coluna contendo
# os 19 grupos de causa e a segunda coluna contendo o numero de casos:
# --------------------------------------------------------------------
tabela <- descreveCSAP(df)
desenhaCSAP(tabela, jaetabela = TRUE)
\dontrun{
# Resulta em erro, faltou o argumento 'quando'
desenhaCSAP(tabela, jaetabela = TRUE, titulo = "auto", onde = 'RS')
}
desenhaCSAP(tabela, jaetabela = TRUE,
            titulo = "auto", onde = "RS", quando = "jan/2012")
desenhaCSAP(tabela, jaetabela = TRUE, titulo = "Título manual")

#  Graficos com as funcoes basicas
# =================================
desenhaCSAP(df, tipo.graf = "base")
desenhaCSAP(df$grupo, tipo.graf = "base")
desenhaCSAP(tabela, jaetabela = TRUE, tipo.graf = "base")

}
\seealso{
\code{\link{csapAIH}}, \code{\link{descreveCSAP}}, \code{\link[ggplot2]{ggplot}}, \code{\link{nomesgruposCSAP}}
}
