% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxetar_quinq.R
\name{fxetar_quinq}
\alias{fxetar_quinq}
\title{Categoriza a idade em faixas etárias quinquenais}
\usage{
fxetar_quinq(aetas = NULL, senectus = 80, puer = FALSE)
}
\arguments{
\item{aetas}{Se NULL (padrão), retorna um vetor com as faixas etárias definidas por \code{puer} e \code{senectus}. Se um vetor com valores numéricos (\code{dbl}, \code{num}, \code{int}), idealmente uma variável com valores de idade, classifica o valor na faixa etária.}

\item{senectus}{Um valor definindo o início do último intervalo, que é aberto. Com o padrão \code{senectus = 80}, a função retorna um fator (\code{fct}) com 17 nívels (\code{levels}) em que o último é "80 e +"}

\item{puer}{Se \code{TRUE}, a primeira faixa etária será quebrada em "< 1 ano" e "1-4"}
}
\description{
Categoriza um vetor de valores contínuos ou inteiros em faixas quinquenais. A primeira faixa pode ser quebrada em "< 1 ano" e "1-4" e a última pode ser definida pelo usuário. Por padrão são 17 faixas quinquenais: 0-4, ..., 80 e + anos de idade.
}
\examples{
data("aih500")
idade <- csapAIH(aih500)$idade
table(fxetar_quinq(idade))
table(fxetar_quinq(idade, senectus = 90, puer = TRUE))

# ou com a função \code{idadeSUS} (e encadeamento/'piping' de comandos):
idadeSUS(aih500)$idade |>
  fxetar_quinq() |>
  table()

}
