% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupnamesCSAP.R
\name{groupnamesCSAP}
\alias{groupnamesCSAP}
\title{Names of Brazilian Primary Health Care Sensitive Conditions groups of cause}
\usage{
groupnamesCSAP(idioma = "pt.ca", nomes = NULL)
}
\arguments{
\item{idioma}{Character vector indicating how to type the names; may be "pt.ca" (default) for names in Portuguese with accents, "pt.sa" for names in Portuguese without accents, "en", for names in English, or "es", for names in Spanish.}

\item{nomes}{Allows a personalized list of group names.}
}
\description{
List the 19 groups of cause from the Brazilian list of Primary Health Care Sensitive Conditions (PHCSC, CSAP).
}
\examples{
groupnamesCSAP()
groupnamesCSAP(idioma = "pt.sa")
groupnamesCSAP(idioma = "en")
groupnamesCSAP(idioma = "es")

}
\references{
#' Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}
}
