% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legoAIH.R
\name{legoAIH}
\alias{legoAIH}
\title{Read the Brazilian Hospital Information System data bases (BD-SIH/SUS)}
\usage{
legoAIH(
  x,
  file = TRUE,
  procobst.rm = TRUE,
  longa.rm = TRUE,
  cep = TRUE,
  cnes = TRUE,
  sep,
  language = "pt",
  ...
)
}
\arguments{
\item{x}{A data file with the Brazilian records for hospital discharges from the Unified Health System ('AIH files -- arquivos da AIH'), in .DBC, .DBF or .CSV format, or yet an object of class data.frame on the active session, with the same structure of AIH files.}

\item{file}{TRUE (default) indicates that \code{x} is a file}

\item{procobst.rm}{TRUE (default) removes records for obstetric procedures hospitalization. See 'Details'.}

\item{longa.rm}{TRUE (default) removes records for large continuance hospitalizations. See 'Details'.}

\item{cep}{TRUE (default) includes the patient ZIP code in the resulting data frame; if the read data frame has the SIHSUS structure but doesn't include that variable, the argument should be passed to FALSE, otherwise the function will be stopped, by error.}

\item{cnes}{TRUE (default) includes the hospital registry on the Brazilian National Register for Health Centers ("Cadastro Nacional de Estabelecimentos de Saúde -- CNES)" in the returned data frame; if the read data frame has the SIHSUS structure but doesn't include the variable, this argument should be passed to FALSE, or the function will be stopped by error.}

\item{sep}{required argument for reading .CSV files, indicating the field separator; possible values are "," and ";".}

\item{language}{Language for displayed messages, summary of reading data, and variable labels; may be "pt" (default) for Portuguese or "en" for English.}

\item{...}{Further arguments passed to \code{\link{read.dbf}}, \code{\link{read.csv}}, and \code{\link[read.dbc]{read.dbc}}.}
}
\description{
Read files from the Brazilian Hospital Information System data bases (BD-SIH/SUS), in .DBC, .DBF or .CSV format, and allows to exclude, extract or identify records of hospitalization for obstetric procedures, as well to exclude duplicated records for large continuance hospitalization ("large stay AIH"). The function also makes changes in variables to be more usable: computes age in completed years, "detailed age groups" (as DATASUS tables) and five-years age groups, turns sex into a factor, and returns a data frame with less (and more used in epidemiological studies) variables.
}
\details{
\itemize{
  \item[procobst.rm] When analyzing the proportion of causes of hospitalization (i.e. the 'proportional morbidity'), whe are usually not interested in counting in the denominator the obstetric admissions which are not due to illness. Thus, the argument \code{procobst.rm} (which equals to TRUE for default) is used to exclude admissions for childbirth or post-abortion curettage.
  \item[longa.rm] In the Unified Health System (SUS), when a patient stays hospitalized for more than a month, the hospital must generate other hospital admission form (AIH) for the same case. These records ("type 5 AIH") are indeed duplicated records, and thus the argument \code{longa.rm} (which for default equals to TRUE) is used to remove or not those records.
 }
}
\examples{
## An 'AIH file' in the working directory
##-----------------------------------------
\dontrun{
aih <- legoAIH('RDSC1201.dbc')
str(aih)
attributes(aih)$Resumo
}

## A data frame with the 'AIH files' structure
##---------------------------------------------
data("aih500")
aih <- legoAIH(aih500)
str(aih)
attributes(aih)$Resumo

}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2014. SIH – Sistema de Informação Hospitalar do SUS: Manual Técnico Operacional do Sistema. Ministério da Saúde: Brasília, 2015. 87p.
}
\seealso{
\code{\link{csapAIH}}, \code{\link{idadeSUS}}, \code{\link{read.csv}}, \code{\link[read.dbc]{read.dbc}} (to read the DATASUS compressed .DBC files), and \code{\link{read.dbf}}
}
