% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leraih.R
\name{leraih}
\alias{leraih}
\title{Ler os arquivos da AIH.}
\usage{
leraih(
  x,
  arquivo = TRUE,
  procobst.rm = TRUE,
  parto.rm = TRUE,
  longa.rm = TRUE,
  vars = NULL,
  ...
)
}
\arguments{
\item{x}{Alvo da função. Banco de dados com a estrutura dos arquivos da AIH, carregado como \code{data frame} na seção de trabalho ou armazenado em arquivo em formato dbc, dbf ou csv.}

\item{arquivo}{x é um arquivo de dados? Padrão é \code{TRUE}.}

\item{procobst.rm}{As internações para procedimentos obstétricos devem ser excluídas? Padrão é \code{TRUE}. V. details.}

\item{parto.rm}{argumento lógico, obrigatório se \code{sihsus=TRUE}; \code{TRUE} (padrão) exclui as internações por parto (\code{ver detalhes});}

\item{longa.rm}{As internações de longa permanência (tipo 5) devem ser excluídas? Padrão é \code{TRUE}. V. details.}

\item{vars}{Variáveis da AIH para seleção. O padrão (\code{vars = NULL}), seleciona o conjunto usado em \code{\link{csapAIH}}. V. details.}

\item{...}{outros parâmetros das funções utilizadas

#' @details
 \itemize{
  \item x pode ser:
   \enumerate{
    \item um arquivo de dados armazenado num diretório;
    \item um banco de dados, ou um vetor da classe \code{factor} presente como objeto no espaço de trabalho do R, em que uma das variáveis, ou o vetor, contenha códigos da CID-10.
   }
Se for um arquivo, o nome deve ser escrito entre aspas e com a extensão do arquivo (DBC, DBF ou CSV, em minúsculas ou maiúsculas). Se não estiver no diretório de trabalho ativo, seu nome deve ser precedido pelo caminho (path) até o diretório de armazenamento. Se estiver em outro formato, podem-se usar os argumentos da função \code{\link{read.table}} para leitura dos dados.

\item \code{procbst.rm} TRUE (padrão) exclui as internações por procedimentos relacionados ao parto ou abortamento. São excluídas as internações pelos seguintes procedimentos obstétricos, independente do diagnóstico principal de internação (variável `DIAGPRINC`):
  \itemize{
   \item 0310010012  ASSISTENCIA AO PARTO S/ DISTOCIA
   \item 0310010020  ATENDIMENTO AO RECÉM-NASCIDO EM SALA DE PARTO
   \item 0310010039  PARTO NORMAL
   \item 0310010047  PARTO NORMAL EM GESTAÇÃO DE ALTO RISCO
   \item 0411010018  DESCOLAMENTO MANUAL DE PLACENTA
   \item 0411010026  PARTO CESARIANO EM GESTAÇÃO ALTO RISCO
   \item 0411010034  PARTO CESARIANO
   \item 0411010042  PARTO CESARIANO C/ LAQUEADURA TUBÁRIA
   \item 0411020013  CURETAGEM PÓS-ABORTAMENTO / PUERPERAL
   \item 0411020021  EMBRIOTOMIA
   }
\item \code{parto.rm } TRUE (padrão) exclui as internações por parto pelo campo diagnóstico, independente do procedimento. São excluídas as internações com os seguintes diagnósticos (CID-10):
 \itemize{
  \item O80 Parto único espontâneo
  \item O81 Parto único por fórceps ou vácuo-extrator
  \item O82 Parto único por cesariana
  \item O83 Outros tipos de parto único assistido
  \item O84 Parto múltiplo
 }

É retornada uma mensagem informando o número de registros lidos, o número e proporção de registros excluídos e o total de registros importados.

}}
}
\description{
Lê os "arquivos reduzidos" da Autorização de Internação Hospitalar (AIH).
}
