% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomesgruposCSAP.R
\name{nomesgruposCSAP}
\alias{nomesgruposCSAP}
\title{Nomes dos grupos de causa da Lista Brasileira de Condições Sensíveis à Atenção Primária}
\usage{
nomesgruposCSAP()
}
\description{
Lista os 19 grupos de causa CSAP, em ordem crescente
}
\author{
Fúlvio B. Nedel
}
