% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partos.R
\name{partos}
\alias{partos}
\title{Identificar diagnósticos de parto na CID-10}
\usage{
partos(cid, opus = 1)
}
\arguments{
\item{cid}{Vetor com o código diagnóstico, segundo a CID-10.}

\item{opus}{Ação a ser executada: 1 (padrão) seleciona apenas os partos; 2 exclui os partos; e 3 cria um \code{data.frame} com duas variáveis -- \code{cid}, com o código diagnóstico, e \code{parto} (sim ou não).}
}
\description{
Identificar diagnósticos de parto na CID-10
}
\examples{
data(aih100)
# Selecionar os diagnósticos de parto em um vetor
partos(aih100$DIAG_PRINC)
# Selecionar os diagnósticos de parto em um vetor (mesma coisa, explicitando o argumento 'opus')
partos(aih100$DIAG_PRINC, 1)
# Excluir os diagnósticos de parto em um vetor
partos(aih100$DIAG_PRINC, 2)
# Criar um 'data frame' com duas variáveis classificando os diagnósticos em parto ou não
partos(aih100$DIAG_PRINC, 3)
#' # Selecionar apenas os registros de partos em um banco de dados
aih100[aih100$DIAG_PRINC == partos(aih100$DIAG_PRINC), ]
#' # Excluir os registros de partos em um banco de dados
aih100[aih100$DIAG_PRINC == partos(aih100$DIAG_PRINC, 2), ]
#' # Acrescentar ao banco uma coluna com a variável "parto"
aih100 <- cbind(aih100, partos(aih100$DIAG_PRINC, 3)[2]) |> print()
aih100$parto2 <- partos(aih100$DIAG_PRINC, 3)$parto

}
