% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCSAP.R
\name{plotCSAP}
\alias{plotCSAP}
\title{Plot Primary Health Care Sensitive Conditions, Brazilian List}
\usage{
plotCSAP(
  x,
  title = NULL,
  where,
  when = NULL,
  t.hjust = 1,
  t.size = 12,
  x.size = 10,
  y.size = 12,
  x.cte = 5,
  breaks = NULL,
  limits = NULL,
  idioma = "pt.ca",
  ...
)
}
\arguments{
\item{x}{Object containing the information to be plotted. May be a factor, a character vector, a table or a data frame with the 19 groups of cause CSAP from the Brazilian List.}

\item{title}{Character vector for graphic title. If not NULL, may be "auto", for an automatic title.}

\item{where}{Character vector indicating the region where people are from, to be passed to automatic title; needed only if title = "auto".}

\item{when}{Character vector indicating the time period of hospitalizations, to be passed to automatic title; needed only if title = "auto" AND user wants other period than the month of hospitalizations, in monthly "AIH files".}

\item{t.hjust}{Numeric value for horizontal adjust of the title.}

\item{t.size}{Numeric value for the title font size.}

\item{x.size}{Numeric value for the font size of x axis label.}

\item{y.size}{Numeric value for the font size of y axis label.}

\item{x.cte}{Numeric value for adjusting the space to the right of bars, and vertical grid spacement, as well to tick marks and labels for x axis; ignored for \code{breaks} and \code{limits} if these arguments are set.}

\item{breaks}{Numeric vector to set the breaks for the horizontal (actually y) axis; if not NULL (default) cause \code{x.cte} to be ignored for setting the breaks.}

\item{limits}{Numeric vector of length 2, to set the space to the right of bars; if not NULL (default) cause \code{x.cte} to be ignored for setting this space.}

\item{idioma}{Mode of CSAP group names; argument to be passed to \code{\link{groupnamesCSAP}}.}

\item{...}{Arguments to be passed to other functions.}
}
\description{
Plot an horizontal bar graphic (using \code{\link[ggplot2]{ggplot2}}) for Primary Health Care Sensitive Conditions (PHCSC, CSAP), according to the Brazilian List of causes.
}
\examples{
data("aih500")
csap <- csapAIH(aih500)
plotCSAP(csap)
plotCSAP(csap, idioma = "en")
plotCSAP(csap$grupo)
plotCSAP(csap$grupo, idioma = "pt.sa")
plotCSAP(csap$grupo, idioma = "en")
plotCSAP(csap$grupo, idioma = "es")

}
\seealso{
\code{\link{csapAIH}}, \code{\link[ggplot2]{ggplot2}} \code{\link{groupnamesCSAP}}, \code{\link{legoAIH}} and \code{\link{tableCSAP}}
}
