% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POPBR10.R
\docType{data}
\name{POPBR10}
\alias{POPBR10}
\title{População dos municípios brasileiros, 2010}
\format{
An object of class \code{data.frame} with 734580 rows and 6 columns.
}
\source{
\url{ftp://ftp.datasus.gov.br/dissemin/publicos/IBGE/POP/POPBR10.zip}
}
\usage{
data(POPBR10)
}
\description{
Banco de dados com a contagem populacional por sexo e faixa etária dos municípios brasileiros, Censo 2010.
}
\examples{
# data("POPBR10")
# str(POPBR10)

}
\keyword{datasets}
