% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popbr2000_2021.R
\name{popbr2000_2021}
\alias{popbr2000_2021}
\title{População Brasil 2000-2021}
\usage{
popbr2000_2021(anoi = NULL, anof = NULL, uf = NULL, munic = NULL)
}
\arguments{
\item{anoi}{Ano de início do período a ser considerado, de 2000 a 2021. Argumento opcional, se não for preenchido, são consideradas as estimativas para todos os anos disponíveis (v. Examples).}

\item{anof}{Ano de fim do período a ser considerado, de 2000 a 2021. Argumento opcional, se não for preenchido, são consideradas as estimativas para todos os anos disponíveis (v. Examples).}

\item{uf}{Vetor com a(s) sigla(s) da(s) Unidade(s) da Federação de interesse.}

\item{munic}{Vetor com o código IBGE do(s) municípios(s) de interesse.}
}
\description{
Estimativas populacionais para os municípios brasileiros por sexo e faixa etária quinquenal, por ano de 2000 a 2021. Modifica a tabela de dados resultante da função \code{\link[brpop]{mun_sex_pop}} do pacote [`brpop`](https://github.com/rfsaldanha/brpop), permitindo seu uso como um banco de dados, além de outras mudanças (v. Details).
}
\details{
Colocar o link pro TABNET, detalhar as mudanças feitas.
}
\examples{
# Ano 2021, todos os municípios brasileiros:
popbr2000_2021(2021)
# Anos 2019 a 2021, RS:
popbr2000_2021(2019, uf = "RS")
# Anos 2000 a 2003, AC:
popbr2000_2021(anof = 2003, uf = "AC")
# Ano 2015, Cerro Largo, RS:
popbr2000_2021(2015, 2015, munic = "430520")


}
\seealso{
\code{\link[brpop]{mun_sex_pop}}
}
