% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procobstb.R
\name{proc.obstb}
\alias{proc.obstb}
\title{Obstetric procedures from the Brazilian Hospital Information System data bases (BD-SIH/SUS)}
\usage{
proc.obstb(
  x,
  procobst.action = "exclude",
  proc.rea = "PROC_REA",
  language = "pt"
)
}
\arguments{
\item{x}{A data frame with the Brazilian records for hospital discharges from Unified Health System, the Hospital Admission Autorization Form -- "Autorização de Internação Hospitalar (AIH)", v.g. "AIH files -- arquivos da AIH", on the Bases de Dados do Sistema de Informacao Hospitalar do SUS -- BD-SIH/SUS.}

\item{procobst.action}{Character argument indicating the action to be fulfilled on the data frame: (1) \code{"exclude"} (default) returns a data frame without the hospitalizations for obstetric procedures, (2) \code{"extract"} returns a data frame with only the hospitalizations for obstetric procedures, (3) \code{"identify"} returns a data frame with all the original records plus one variable (\code{procobst}) of class \code{factor} indicating whether the hospitalization was for an obstetric procedure or not.}

\item{proc.rea}{Procedure performed, name to the "procedimento realizado" field (\code{PROC_REA}) on the AIH file).}

\item{language}{Language for displayed messages and summary of fulfilled actions; may be "pt" (default) for portuguese or "en" for english.}
}
\description{
Identifies records of hospitalizations for non-morbidity obstetric procedures (v.g. deliveries etc.) from the Brazilian Hospital Information System data bases (BD-SIH/SUS) and, according to user options, (1) create a new variable \code{procobst} with the procedure identification, (2) exclude these records from the data frame, or (3) create a new data frame with only these records.
}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2014. SIH – Sistema de Informação Hospitalar do SUS: Manual Técnico Operacional do Sistema. Ministério da Saúde: Brasília, 2015. 87p.
}
\seealso{
\code{\link{legoAIH}} and \link{csapAIH}
}
