% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabCSAP.R
\name{tabCSAP}
\alias{tabCSAP}
\title{Tabulate Primary Health Care Sensitive Conditions (PHCSC)}
\usage{
tabCSAP(x, digits = 2, tipo = "pt.ca", format = FALSE)
}
\arguments{
\item{x}{a vector containing the Brazilian PHCSC groups of causes.}

\item{digits}{number of decimals to be rounded (with \code{\link{round}}).}

\item{tipo}{argument to be passed to \code{\link{nomesgruposCSAP}}.}

\item{format}{Should the table be formatted to print? (default = FALSE).}
}
\description{
Tabulates Primary Health Care Sensitive Conditions (PHCSC, CSAP) according to the Brazilian list of causes.
}
\examples{
data("aih500")
csap <- csapAIH(aih500)
tabCSAP(csap$grupo)
# tabCSAP(csap$grupo, tipo = "pt.sa")
# tabCSAP(csap$grupo, tipo = "en")
# tabCSAP(csap$grupo, tipo = "es")

}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}
}
\seealso{
\code{\link{csapAIH}}, \code{\link{nomesgruposCSAP}}
}
