% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCSAP.R
\name{tableCSAP}
\alias{tableCSAP}
\alias{csapAIH}
\title{Tabulate Primary Health Care Sensitive Conditions (PHCSC)}
\usage{
tableCSAP(x, idioma = "pt.ca")
}
\arguments{
\item{x}{a vector containing the Brazilian PHCSC groups of causes.}

\item{idioma}{argument to be passed to \code{\link{groupnamesCSAP}}.}
}
\description{
Tabulates Primary Health Care Sensitive Conditions (PHCSC, CSAP) according to the Brazilian list of causes.
}
\examples{
data("aih500")
csap <- csapAIH(aih500)
tableCSAP(csap$grupo)
tableCSAP(csap$grupo, idioma = "pt.sa")
tableCSAP(csap$grupo, idioma = "en")
tableCSAP(csap$grupo, idioma = "es")

}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}
}
\seealso{
\code{\link{csapAIH}}, \code{\link{groupnamesCSAP}}, \code{\link{legoAIH}}
}
