% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufbr.R
\name{ufbr}
\alias{ufbr}
\title{Lista as Unidades da Federação por Região do Brasil}
\usage{
ufbr(nomes = FALSE)
}
\arguments{
\item{nomes}{Se \code{nomes = TRUE}, é incluída uma coluna com os nomes dos Estados. Se \code{nomes = FALSE} (padrão), a UF é identificada apenas pela sigla.}
}
\value{
Um objeto da classe \code{data.frame}, com a sigla da Região e o código, a sigla e, se \code{nomes = TRUE}, o nome da UF.
}
\description{
Lista as siglas e nomes das Unidades Federativas (UF) por Região do Brasil segundo seu código no IBGE, permitindo a decodificação do código do IBGE para a sigla ou nome da Unidade da Federação.
}
\examples{
ufbr()
ufbr(nomes = TRUE)
# Para acrescentar as informações a um banco de dados existente, junte os
# bancos, como no exemplo abaixo.
# O banco "POPBR10" tem a população dos municípios brasileiros em 2010.
# O município é informado pelo código do IBGE e não há informação sobre a UF:
data("POPBR10")
str(POPBR10)
# O código abaixo
# - acrescenta a 'POPBR10' uma variável chamada "CO_UF" com o código da UF,
# tomado dos dois primeiros dígitos de 'MUNIC_RES';
# - une os bancos, com a função \code{\link{merge}}; e
# - seleciona os registros da Região Norte ("N")

POPBR10$CO_UF <- substr(POPBR10$MUNIC_RES, 1, 2)
POPBR10  <- merge(POPBR10, ufbr())
POPBR10[POPBR10$REGIAO == "N", ]
}
