% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomesgruposCSAP.R
\name{adinomes}
\alias{adinomes}
\title{Adiciona o nome dos grupos}
\usage{
adinomes(x, lista = "MS")
}
\arguments{
\item{x}{Banco de dados}

\item{lista}{Lista CSAP a ser utilizada. O padrão é "MS" (v. \link{nomesgruposCSAP}).}
}
\description{
Acrescenta uma variável com o nome dos grupos de causa segundo a lista selecionada ("MS" ou "Alfradique") a um banco de dados resultante da função `csapAIH` ou que contenha uma variável de nome "grupo" com os grupos nomeados segundo aquela função ("g01", ...)
}
\details{
Define como missing os não-CSAP
}
\examples{
data("aih100")
adinomes(csapAIH(aih100))[1:5, 9:11]
}
\seealso{
nomesgruposCSAP()

[listaBRMS()]

[listaBRAlfradique]
}
