% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aih100.R
\docType{data}
\name{aih100}
\alias{aih100}
\title{Banco com 100 registros de um 'arquivo da AIH' (RD??????.DBC).}
\format{
An object of class \code{data.frame} with 100 rows and 113 columns.
}
\source{
\url{http://www2.datasus.gov.br/DATASUS/index.php?area=0901}
}
\usage{
data(aih100)
}
\description{
Banco de dados (classe \code{data.frame}) com 100 observações de um "arquivo da AIH", 2018.
}
\examples{
data("aih100")
str(aih100)

}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2014. SIH – Sistema de Informação Hospitalar do SUS: Manual Técnico Operacional do Sistema. Ministério da Saúde: Brasília, 2015. 87p.
}
\keyword{datasets}
