% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aih500.R
\docType{data}
\name{aih500}
\alias{aih500}
\title{Banco com 500 registros de um "arquivo da AIH", ano 2018}
\format{
An object of class \code{data.frame} with 500 rows and 113 columns.
}
\source{
\url{http://www2.datasus.gov.br/DATASUS/index.php?area=0901}
}
\usage{
data(aih500)
}
\description{
Amostra aleatória de 500 registros e todas as variáveis de um "arquivo da AIH" (RD??????.DBC), RS, janeiro de 2018.
}
\examples{
data("aih500")
str(aih500)
}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Departamento de Regulação, Avaliação e Controle. Coordenação Geral de Sistemas de Informação - 2014. SIH – Sistema de Informação Hospitalar do SUS: Manual Técnico Operacional do Sistema. Ministério da Saúde: Brasília, 2015. 87p.
}
\keyword{datasets}
