% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cid10cap.R
\name{cid10cap}
\alias{cid10cap}
\title{Função para classificar códigos da CID-10 em Capítulos}
\usage{
cid10cap(cid, droplevels = FALSE)
}
\arguments{
\item{cid}{Nome do vetor com os códigos da CID-10}

\item{droplevels}{Se TRUE, desconsidera os capítulos sem nenhuma ocorrência de casos. O padrão é FALSE, o que retorna uma tabela com zeros nos capítulos sem ocorrência de casos.}
}
\description{
Tomando um vetor com códigos da CID-10, acrescenta uma variável com os capítulos da CID-10 correspondentes
}
\examples{
# Arquivos da AIH
cid10cap(aih500$DIAG_PRINC) |> table()
cid10cap(aih500$DIAG_PRINC, droplevels = TRUE) |> table()

# Arquivos da DO
\dontrun{
  cid10cap(Rcoisas::obitosRS2019$CAUSABAS) |> table()
}

}
