% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cid10cap.R
\name{cid10cap}
\alias{cid10cap}
\title{Função para classificar códigos da CID-10 em Capítulos}
\usage{
cid10cap(cid, data)
}
\arguments{
\item{cid}{nome da variável com os códigos da CID-10}

\item{data}{nome do banco de dados}
}
\description{
Tomando um banco de dados com uma variável com códigos da CID-10, acrescenta uma variável com os capítulos da CID-10 correspondentes
}
\examples{
\dontrun{
 require(microdatasus)
 DORS19 <- fetch_datasus(2019, 01, 2019, 12, "RS", "SIM-DO")
 DORS19 <- cid10cap("CAUSABAS", DORS19)
 tabuleiro(DORS19$capcid)
 par(mar= c(5, 15, 1, 1))
 barplot(rev(table(DORS19$capcid)), horiz = T, las = 1, cex.names = .7, )
}

}
