% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csapAIH-internal.R
\docType{package}
\name{csapAIH-package}
\alias{csapAIH-package}
\alias{_PACKAGE}
\title{Pacote csapAIH}
\description{
Classifica um vetor com códigos da CID-10 segundo a Lista Brasileira de Condições Sensíveis à Atenção Primária e oferece outras funcionalidades, especialmente para o manejo dos 'arquivos da AIH' (BD-SIH/SUS).

A principal função do pacote, \code{\link{csapAIH}}, classifica os códigos da CID-10 segundo a Lista Brasileira de CSAP. \code{\link{descreveCSAP}} constrói uma tabela com frequências absolutas e relativas dos casos por grupos de causa. \code{\link{desenhaCSAP}} constrói um gráfico de barras por grupos de causa. \code{\link{idadeSUS}} calcula a idade do paciente na internação, nos "arquivos da AIH".
}
\details{
\itemize{
   \item \code{v0.0.3.2}

     Criada a função \code{\link{proc.obst}}, para a identificação dos procedimentos obstétricos pela tabela do SIH/SUS, com três possíveis resultados: (1) exclusão dos registros de procedimento obstétrico (padrão); (2) criação de nova variável com identificação (sim/não) do caso; e (3) exclusão dos demais registros. A exclusão dos procedimentos obstétricos (argumento \code{procobst.rm = TRUE}) na função \code{\link{csapAIH}} agora é feita invocando a função \code{\link{proc.obst}}. O argumento \code{parto.rm} agora funciona em data frames sem a estrutura do SIH/SUS. Algumas melhoras na documentação do pacote.
   \item \code{v0.0.3.1}
     \itemize{
       \item a função \code{\link{csapAIH}} passa a ter um argumento \code{cid = NULL}, para identificar colunas com o CID em um \code{data.frame};
       \item a função \code{\link{desenhaCSAP}} recebe os argumentos \code{val.dig = 0}, para definir o nº de decimais nas porcentagens apresentadas nas barras do gráfico, e \code{val.size = 2.5}, para definir o tamanho das letras dos valores das barras;
       \item criado o argumento \code{sis} com as opções \code{"SIH"} (padrão) e \code{"SIH"} em \code{\link{idadeSUS}}; finalmente a função pode ser utilizada para computar a idade nas bases da Declaração de Óbito no Sistema de Informações sobre Mortalidade
         }

   \item \code{v0.0.3}

   As funções acrescentadas na versão 0.0.2 foram ampliadas: podem ser tabulados vetores sem todos os grupos de causa; grupos com frequência zero são excluídos do gráfico; o gráfico pode ser gerado com funções básicas, por demanda do usuário ou se \code{\link[ggplot2]{ggplot2}} não estiver instalado no sistema; quando gerado a partir de um banco de dados, o gráfico em \code{\link[ggplot2]{ggplot}} pode ser estratificado segundo categorias de outras variáveis do banco. O cálculo da idade não é mais feito na função \code{\link{csapAIH}}, mas por uma nova função, \code{\link{idadeSUS}}, invocada por \code{\link{csapAIH}} e que também pode ser chamada independentemente sobre um "arquivo da AIH".

   \item \code{v0.0.2}

   Foram acrescentadas três funções ao pacote: \code{\link{descreveCSAP}}, \code{\link{desenhaCSAP}} e \code{\link{nomesgruposCSAP}}, para a realização de tabelas e gráficos com qualidade de impressão para artigos e apresentação de relatórios. Acrescentada dependência do pacote \code{\link[read.dbc]{read.dbc}} para a leitura de arquivos comprimidos do DATASUS, em formato DBC.

   \item \code{v0.0.1}

   O pacote tem apenas uma função, \code{csapAIH()}, que permite trabalhar com um objeto no espaço de trabalho ou ler um arquivo com os dados. Se os dados tiverem a estrutura dos "arquivos da AIH" com uma definição mínima de variáveis, a função, com suas opções padrão, além de classificar a internação em CSAP, computa a idade em anos completos e em duas classificações etárias e permite a exclusão das AIHs de "longa permanência" e internações por alguns procedimentos obstétricos definidos.
   }
}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No. 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}

Nedel FB. csapAIH: uma função para a classificação das condições sensíveis à atenção primária no programa estatístico R. Epidemiol. Serv. Saúde. 2017; 26(1):199-209. Disponível em: \cr \url{http://scielo.iec.gov.br/scielo.php?script=sci_arttext&pid=S1679-49742017000100199&lng=pt}.
}
\keyword{ACSC,}
\keyword{AIH-SUS}
\keyword{CSAP,}
\keyword{PHCSC}
\keyword{package}
