\name{csapAIH-package}
\alias{csapAIH-package}
%\alias{csapAIH}
\docType{package}
\title{
\packageTitle{csapAIH}
}
\description{
% \packageDescription{csapAIH}
Classifica códigos da CID-10 segundo a Lista Brasileira de Condições Sensíveis à Atenção Primária e oferece outras funcionalidades, especialmente para o manejo dos "arquivos da AIH" (RD??????.DBF; BD-SIH/SUS).

O pacote tem apenas uma função, \code{csapAIH()}, que permite trabalhar com um objeto no espaço de trabalho ou ler um arquivo com os dados. Se os dados tiverem a estrutura dos "arquivos da AIH" com uma definição mínima de variáveis, é computada a idade em anos completos e em duas classificações etárias, são excluídas as AIHs de "longa permanência" e podem ser excluídas internações por alguns procedimentos obstétricos definidos. Veja a ajuda da função para detalhes.
}
\details{
\packageDESCRIPTION{csapAIH}
\packageIndices{csapAIH}
%%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
% \packageAuthor{csapAIH}
Fúlvio B. Nedel
Maintainer: Fúlvio B. Nedel % \packageMaintainer{csapAIH}
}
%\references{
%%~~ Literature or other references for background information ~~
%}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%%~~ simple examples of the most important functions ~~
%}
