% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descreveCSAP.R
\name{descreveCSAP}
\alias{descreveCSAP}
\title{Tabular Condicoes Sensiveis a Atencao Primaria}
\usage{
descreveCSAP(grupos, digits = 2)
}
\arguments{
\item{grupos}{Pode ser:
\itemize{
  \item Um \code{data.frame} gerado pela função \code{\link{csapAIH}}, ou qualquer \code{data.frame} com uma variável chamada \code{grupo} com os grupos de causa da Lista Brasileira de CSAP, rotulados como os resultantes da função \code{\link{csapAIH}}, isto é, "g01", "g02", ..., "g19".
  \item Um vetor da classe \code{factor} ou \code{character} com os grupos de causa CSAP, nomeados de acordo com o resultado da função \code{\link{csapAIH}}. Esse vetor não precisa ser gerado pela função \code{\link{csapAIH}}, mas deve conter todos os 19 grupos de causa, ainda que sua frequência seja zero, e também devem ser rotulados da mesma forma e ordem que na função, isto é, "g01", "g02", ..., "g19".
  }}

\item{digits}{Número de decimais nas proporções apresentadas.}
}
\description{
Constrói uma tabela de frequências absolutas e relativas das CSAP por grupo de causa
}
\examples{
data(aih100)
df = csapAIH(aih100)
descreveCSAP(df$grupo)

}
\seealso{
\code{\link{csapAIH}}, \code{\link{desenhaCSAP}}, \code{\link{nomesgruposCSAP}}
}
