% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descreveCSAP.R
\name{descreveCSAP}
\alias{descreveCSAP}
\title{Tabular Condicoes Sensiveis a Atencao Primaria}
\usage{
descreveCSAP(grupos, digits = 2)
}
\arguments{
\item{grupos}{Vetor com os grupos de causa CSAP, de acordo com o resultado da função \code{\link{csapAIH}}.}

\item{digits}{Número de decimais nas proporções apresentadas.}
}
\description{
Constrói uma tabela de frequências absolutas e relativas das CSAP por grupo de causa
}
\details{
\code{grupos} não precisa ser gerado com a função \code{\link{csapAIH}}, mas deve usar os mesmos caracteres de identificação dos grupos CSAP que o resultado da função, v.g. "g01", "g02", ..., "g19".
}
\examples{
data(aih100)
df = csapAIH(aih100)
descreveCSAP(df$grupo)

}
\seealso{
\code{\link{csapAIH}}, \code{\link{desenhaCSAP}}, \code{\link{nomesgruposCSAP}}
}
\author{
Fúlvio B. Nedel
}
