% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desenhaCSAP.R
\name{desenhaCSAP}
\alias{desenhaCSAP}
\title{Grafico das Condicoes Sensiveis a Atencao Primaria}
\usage{
desenhaCSAP(banco, titulo = NULL, onde, quando = NULL, t.hjust = 1,
  t.size = 12, x.size = 10, y.size = 12, limsup = NULL,
  tipo = "ggplot")
}
\arguments{
\item{banco}{Banco de dados com as informações. Objeto da classe \code{data.frame} gerado pela função \code{\link{csapAIH}}.}

\item{titulo}{Título do gráfico; se NULL (default), não é gerado um título; se \code{"auto"}, o argumento \code{onde} passa a ser obrigatório e a função gera um título para o gráfico a partir da informação de \code{onde} e do arquivo de dados ou do informado para o argumento \code{quando}.}

\item{onde}{Local, população de origem dos dados do gráfico; obrigatório se \code{titulo = "auto"}.}

\item{quando}{Período de referência dos dados; se a fonte de dados for um "arquivo da AIH" (RD????.dbc), pode ser extraído do arquivo.}

\item{t.hjust}{Valor para definição de ajuste horizontal do título. Default é 1.}

\item{t.size}{Valor para definição do tamanho de letra do título. Default é 12.}

\item{x.size}{Tamanho da letra do eixo x. Default é 10.}

\item{y.size}{Tamanho da letra do eixo y. Default é 12.}

\item{limsup}{Valor para ajuste do espaçamento do eixo de frequências.}

\item{tipo}{"ggplot" (padrão) cria um gráfico com \code{\link[ggplot2]{ggplot2}}; quando definido como "base", desenha um gráfico com as funções básicas. Permite que a função desenhe um gráfico mesmo quando \code{\link[ggplot2]{ggplot2}} não está instalado.}
}
\value{
Na opção padrão e com \code{\link[ggplot2]{ggplot2}} instalado, devolve um objeto das classes "gg", "ggplot", com o gráfico.
}
\description{
Desenha um gráfico de barras das CSAP por grupo de causa
}
\details{
O gráfico é desenhado com \code{\link[ggplot2]{ggplot2}}. Portanto, segue essa filosofia e permite adição de outros comandos ao objeto devolvido. \code{grupos} não precisa ser gerado com a função \code{\link{csapAIH}}, mas deve usar os mesmos caracteres de identificação dos grupos CSAP que o resultado da função, v.g. "g01", "g02", ..., "g19".
}
\examples{

data("aih100")
df = csapAIH(aih100)
desenhaCSAP(df)

}
\seealso{
\code{\link{csapAIH}}, \code{\link{descreveCSAP}}, \code{\link[ggplot2]{ggplot2}}, \code{\link{nomesgruposCSAP}}
}
\author{
Fúlvio B. Nedel
}
