% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeh20.R
\docType{data}
\name{eeh20}
\alias{eeh20}
\title{Banco com 1000 registros de hospitalização do Equador.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 36 columns.
}
\source{
\url{https://www.ecuadorencifras.gob.ec/}
O Equador tem mudado os links e nomes ou estrutura da árvore de arquivos. Além disso as últimas buscas têm direcionado a uma página para download dos arquivos em que não se pode copiar o link da fonte, obrigando ao download e salvamento do arquivo em alguma pasta. ... :(
Nas bases disponíveis através da página https://anda.inec.gob.ec/anda/index.php/catalog/SOCDEMO parece que ainda podemos copiar o link do arquivo-fonte desejado. O deste exemplo, a _Encuesta de Egresos Hospitalarios_ do ano 2020, é, neste momento, esse: https://anda.inec.gob.ec/anda/index.php/catalog/883/get_microdata.
}
\usage{
data(eeh20)
}
\description{
A *Encuesta de Egresos Hospiatalarios* registra as hospitalizações ocorridas em todo o Equador, no setor público e privado. Aqui é apresentada uma amostra aleatória da base de dados do ano 2020.
}
\examples{
data("eeh20")
str(eeh20)

}
\references{
Arrumar
}
\keyword{datasets}
