% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchcsap.R
\name{fetchcsap}
\alias{fetchcsap}
\title{Baixa os arquivos da AIH e classifica em CSAP}
\usage{
fetchcsap(
  uf = "all",
  anoinicio,
  mesinicio = 1,
  mesfim = mesinicio,
  anofim = anoinicio,
  ...
)
}
\arguments{
\item{uf}{Unidade da Federação. A sigla da UF ou um vetor com as siglas das UF de interesse, entre aspas e em letras maiúsculas. Para todo o Brasil (padrão), use "all".}

\item{anoinicio}{Ano de competência da AIH para início da seleção dos dados, em formato numérico; sem padrão}

\item{mesinicio}{Mês de competência da AIH para início da seleção dos dados, em formato numérico; por padrão é 1}

\item{mesfim}{Mês de competência da AIH para fim da seleção dos dados, em formato numérico; por padrão é igual ao mês de início}

\item{anofim}{Ano de competência da AIH para fim da seleção dos dados, em formato numérico; por padrão é igual ao ano de início}

\item{...}{Permite o uso de parâmetros de \code{\link{csapAIH}}}
}
\value{
Um objeto das classes data.table e data.frame com as seguintes variáveis:
\describe{
  \item{\code{munres}}{Município de residência do paciente}
  \item{\code{munint}}{Município de internação do paciente}
  \item{\code{sexo}}{Sexo do paciente}
  \item{\code{idade}}{Idade do paciente em anos completos}
  \item{\code{fxetar.det}}{Faixa etária detalhada}
  \item{\code{fxetar5}}{Faixa etária quinquenal (0-4, ..., 76-79, 80 e +)}
  \item{\code{csap}}{Internação por CSAP (sim/não)}
  \item{\code{grupo}}{Grupo de causa da Lista Brasileira de ICSAP, ou "não-CSAP"}
  \item{\code{cid}}{Diagnóstico principal da internação, segundo a Classificação Internacional de Doenças, 10ª Revisão}
  \item{\code{data.inter}}{Data da internação}
  \item{\code{data.saida}}{Data da alta}
}
}
\description{
Descarrega os "arquivos da AIH" (arquivos RD<UFAAMM>.DBC) do site FTP do DATASUS e classifica as internações segundo a Lista Brasileira de Condições Sensíveis à atenção Primária.
}
\details{
\code{fetchcsap} é apenas uma abreviatura para um uso específico da função \code{fetch_datasus}, do pacote \code{microdatasus}, de Raphael Saldanha.
Funciona apenas com o SIH/SUS, através do argumento \code{information_system = "SIH-RD"}
Faz apenas o download das variáveis exigidas pela função \code{\link{csapAIH}}, i.e.: `DIAG_PRINC, NASC, DT_INTER, DT_SAIDA, IDADE, COD_IDADE, MUNIC_RES, MUNIC_MOV, SEXO, N_AIH, PROC_REA, IDENT, CEP, CNES`
}
\examples{
# Internações de todo o Brasil, "mês de competência" janeiro de 2023
\dontrun{
  fetchcsap(anoinicio = 2023)
}
# Internações no RS, jan 2023
fetchcsap("RS", 2023)
## RS, 2023
\dontrun{
fetchcsap("RS", 2023, mesfim = 12)
}

}
