% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchcsap.R
\name{fetchcsap}
\alias{fetchcsap}
\title{Baixa os arquivos da AIH e classifica as internações em CSAP}
\usage{
fetchcsap(
  uf = "all",
  anoinicio,
  mesinicio = 1,
  anofim = anoinicio + 1,
  mesfim = 6,
  periodo = "interna",
  fim.periodo = NULL,
  cep = FALSE,
  cnes = FALSE,
  ...
)
}
\arguments{
\item{uf}{Unidade da Federação. A sigla da UF ou um vetor com as siglas das UF de interesse, entre aspas e em letras maiúsculas. Para todo o Brasil (padrão), use "all".}

\item{anoinicio}{Ano de competência da AIH para início da seleção dos dados, em formato numérico; sem padrão.}

\item{mesinicio}{Mês de competência da AIH para início da seleção dos dados, em formato numérico; por padrão é 1.}

\item{anofim}{Ano de competência da AIH para fim da seleção dos dados, em formato numérico; por padrão é igual ao ano seguinte ao ano de início (\code{anoinicio + 1}).}

\item{mesfim}{Mês de competência da AIH para fim da seleção dos dados, em formato numérico; por padrão é 6 (junho). V. detalhes.}

\item{periodo}{O período definido refere-se ao mês e ano de "competência" da AIH ou à data de internação? O padrão (\code{"interna"}) é a internação. V. detahes.}

\item{fim.periodo}{Se o período for definido pela data de internação, qual a data final? O padrão é 31 de dezembro do ano anterior ao definido no argumênto \code{anofim}. V. detalhes.}

\item{cep}{CEP de internação}

\item{cnes}{CNES do estabelecimento que gerou a AIH}

\item{...}{Permite o uso de outros parâmetros de \code{\link{csapAIH}}}
}
\value{
Um objeto de classes \code{data.table} e \code{data.frame} com as seguintes variáveis:
 \itemize{
  \item \code{munres} Município de residência do paciente
  \item \code{munint} Município de internação do paciente
  \item \code{sexo} Sexo do paciente
  \item \code{idade} Idade do paciente em anos completos
  \item \code{fxetar5} Faixa etária quinquenal (0-4, ..., 76-79, 80 e +)
  \item \code{csap} Internação por CSAP (sim/não)
  \item \code{grupo} Grupo de causa da Lista Brasileira de ICSAP, ou "não-CSAP"
  \item \code{cid} Diagnóstico principal da internação, segundo a Classificação Internacional de Doenças, 10ª Revisão
  \item \code{data.inter} Data da internação
  \item \code{data.saida} Data da alta
 }
}
\description{
Descarrega os "arquivos da AIH" (arquivos RD<UFAAMM>.DBC das Bases de Dados do Sistema de Informações Hospitalares do SUS - BD-SIH/SUS) do site FTP do DATASUS e classifica as internações segundo a Lista Brasileira de Condições Sensíveis à Atenção Primária.
}
\details{
Período de download dos arquivos e de internação dos sujeitos.
\itemize{
  \item Os "arquivos da AIH" são definidos por mês e ano de "competência", e não da data de internação. Assim, o arquivo de um determinado "mês de competência" pode incluir registros de internações ocorridas em outro mês ou ano, enquanto pode não incluir todos os casos ocorridos naquele mês. Por padrão, \code{fetchcsap} usa o argumento \code{periodo = "interna"} para selecionar os casos por data de internação de acordo com o período definido nos argumentos \code{anoinicio}, \code{mesinicio} e \code{anofim}, de modo a iniciar no primeiro dia do ano e mês de competência (\code{anoinicio} e \code{mesinicio}) e terminar em 31 de dezembro do ano anterior ao definido em \code{anofim}.
  \item Assim, por padrão, a função exige apenas a definição do ano de início dos casos. Se o usuário definir apenas esse argumento, \code{fetchcsap} fará o download e leitura dos arquivos (de todo o Brasil) de todos os meses até junho do ano seguinte para então selecionar as internações ocorridas no ano definido em \code{anoinicio}.
}

\code{fetchcsap} é apenas uma abreviatura para um uso específico da função \code{fetch_datasus}, do pacote \code{microdatasus}, de Raphael Saldanha. Funciona apenas com o SIH/SUS, através do argumento \code{information_system = "SIH-RD"}, e faz apenas o download das variáveis exigidas pela função \code{csapAIH}, i.e., \code{DIAG_PRINC, NASC, DT_INTER, DT_SAIDA, IDADE, COD_IDADE, MUNIC_RES, MUNIC_MOV, SEXO, N_AIH, PROC_REA, IDENT, CEP, CNES}.
}
\examples{
# Internações de todo o Brasil, ocorridas no ano de 2023 e registradas até jun/2024:
# Colocando apenas o ano, único argumento obrigatório, a função executa o
# download dos arquivos RD??????.DBC de todas as UF de jan/2023 a jun/2024 e
# então extrais apenas os registros com data de internação em 2023.
\dontrun{
  fetchcsap(anoinicio = 2023)
}
# Diferença entre o mês e ano de "competência" da AIH e a data de internação da pessoa,
# exemplo com as internações no Acre, registradas no mês de competência jan 2023:
ac.comp <- fetchcsap("AC", 2023, mesfim = 1, periodo = 'competencia')
nrow(ac.comp)
summary(ac.comp$data.inter)
ac.int <- fetchcsap("AC", 2023, mesfim = 1)
nrow(ac.int)
summary(ac.int$data.inter)
# Assim, há
nrow(ac.comp) - nrow(ac.int)
# internações registradas naquele mês de competência, mas que ocorreram em outro momento.

}
\seealso{
\code{\link{csapAIH}}
}
