% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchcsap.R
\name{fetchcsap}
\alias{fetchcsap}
\title{Baixa os arquivos da AIH e classifica em CSAP}
\usage{
fetchcsap(uf, mesinicio = 1, anoinicio = NULL, mesfim = 6, anofim = NULL)
}
\arguments{
\item{uf}{Unidade da Federação}

\item{mesinicio}{Mês de competência da AIH para início da seleção dos dados}

\item{anoinicio}{Ano de competência da AIH para início da seleção dos dados}

\item{mesfim}{Mês de competência da AIH para fim da seleção dos dados}

\item{anofim}{Ano de competência da AIH para fim da seleção dos dados}
}
\value{
Um objeto da classe data.table/data.frame com as seguintes variáveis:
\describe{
  \item{\code{munres}}{Município de residência do paciente}
  \item{\code{munint}}{Município de internação do paciente}
  \item{\code{sexo}}{Sexo do paciente}
  \item{\code{idade}}{Idade do paciente em anos completos}
  \item{\code{fxetar.det}}{Faixa etária detalhada}
  \item{\code{fxetar5}}{Faixa etária quinquenal (0-4, ..., 76-79, 80 e +)}
  \item{\code{csap}}{Internação por CSAP (sim/não)}
  \item{\code{grupo}}{Grupo de causa da Lista Brasileira de ICSAP, ou "não-CSAP"}
  \item{\code{cid}}{Diagnóstico principal da internação, segundo a Classificação Internacional de Doenças, 10ª Revisão}
  \item{\code{data.inter}}{Data da internação}
  \item{\code{data.saida}}{Data da alta}
}
}
\description{
Descarrega os "arquivos da AIH" (arquivos RD<UFAAMM>.DBC) do site FTP do DATASUS e classifica as internações segundo a Lista Brasileira de Condições Sensíveis à atenção Primária.
}
\examples{
fetchcsap("RS", anoinicio = 2022, mesfim = 1) |>
  str()
# fetchcsap("RS", anoinicio = 2022, mesfim = 1, anofim = 2022) |>
#   str()

}
