% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxetar3g.R
\name{fxetar3g}
\alias{fxetar3g}
\title{# Faixa etária em três grandes categorias}
\usage{
fxetar3g(idade = NULL, fxetar5 = NULL)
}
\arguments{
\item{idade}{Vetor numérico representando a idade em valores contínuos ou inteiros.}

\item{fxetar5}{Vetor da classe \code{factor} ou \code{character} representando faixas etárias quinquenais, rotuladas conforme o resultado da função \code{\link{fxetar_quinq}} (números separados por hífen, sem espaços).}
}
\description{
Recodifica vetores com a idade em números inteiros ou em faixas etárias quinquenais e três grandes faixas etárias: 0-14, 15-59 e 60 e + anos.
}
\examples{
# Criar vetores para idade e faixa etária quinquenal
idade <- as.integer(runif(100, 0, 100))
fxetar <- fxetar_quinq(idade)
# Computar a faixa etária em três grandes gupos
fxetar3g(idade) |> table()
fxetar3g(fxetar5 = fxetar) |> table()

}
