% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_popbr.R
\name{ler_popbr}
\alias{ler_popbr}
\title{Ler arquivos POPBR????.csv/DBF}
\usage{
ler_popbr(arquivo = NULL, ano = NULL)
}
\arguments{
\item{arquivo}{Nome do arquivo DBF armazenado no computador. Padrão é \code{NULL}, v. details.}

\item{ano}{Ano da estimativa ou contagem populacional a ser capturada no site FTP DATASUS. Padrão é \code{NULL}, v. details.}
}
\description{
Lê os arquivos de população do DATASUS e cria uma variável com a faixa etária quinquenal
}
\details{
Apenas um dos dois parâmetros deve ser preenchido. Se a
}
\examples{
\dontrun{
# Um arquivo no mesmo diretório de trabalho da sessão ativa:
popBR2010 <- ler_popbr("data-raw/POPBR10.DBF")
head(popBR2010)
xtabs(populacao ~ fxetar5 + sexo, data = popBR2010)
}

# Um arquivo no diretório FTP do DATASUS
popBR2010 <- ler_popbr(ano = 2010)
xtabs(populacao ~ fxetar5 + sexo, data = popBR2010)

}
