% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_popbr.R
\name{ler_popbr}
\alias{ler_popbr}
\title{Ler arquivos de população do DATASUS}
\usage{
ler_popbr(x)
}
\arguments{
\item{x}{Nome do arquivo armazenado no computador, ou ano da estimativa ou contagem populacional a ser capturada no site FTP DATASUS. Se o alvo é um arquivo no computador, o nome com a extensão (dbf) deve vir entre aspas. Se o alvo é um arquivo do servidor FTP do DATASUS, deve-se usar o argumento \code{ano}, com o ano (sem aspas) desejado, de 1980 a 2024. Apenas arquivos em formato DBF são lidos.}
}
\description{
Lê os arquivos com estimativas e contagens da população dos municípios brasileiros por sexo e faixa etária disponibilizados pelo DATASUS e entrega um banco de dados com as variáveis originais mais a faixa etária quinquenal.
}
\details{
Essas informações podem ser tabuladas em
 http://tabnet.datasus.gov.br/cgi/deftohtm.exe?ibge/cnv/popsvs2024br.def
}
\examples{
\dontrun{
# Um arquivo no computador:
popBR2010 <- ler_popbr("data-raw/POPBR10.DBF")
head(popBR2010)
xtabs(populacao ~ fxetar5 + sexo, data = popBR2010)

popBR2024 <- ler_popbr("data-raw/POPBR24.DBF")
head(popBR2024)
xtabs(populacao ~ fxetar5 + sexo, data = popBR2024)
}

# Um arquivo no diretório FTP do DATASUS
popBR2010 <- ler_popbr(2010)
head(popBR2010)
xtabs(populacao ~ fxetar5 + sexo, data = popBR2010)

popBR2013 <- ler_popbr(2013)
head(popBR2013)
xtabs(populacao ~ fxetar5 + sexo, data = popBR2013)

popBR2024 <- ler_popbr(2024)
head(popBR2024)
xtabs(populacao ~ fxetar5 + sexo, data = popBR2024)

}
\references{
http://tabnet.datasus.gov.br/cgi/IBGE/SEI_MS-0034745983-Nota_Tecnica_final.pdf
}
