% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_popbr.R
\name{ler_popbr}
\alias{ler_popbr}
\title{Ler arquivos POPBR????.CSV/DBF}
\usage{
ler_popbr(pop, source = "file")
}
\arguments{
\item{pop}{População. Nome do arquivo a ser lido.}

\item{source}{Fonte de dados. Use `file` (padrão) para um arquivo no computador ou `url` para baixar o arquivo do FTP do DATASUS.}
}
\description{
Lê os arquivos de população do DATASUS e cria uma variável com a faixa etária quinquenal
}
\examples{
\dontrun{
# Um arquivo no mesmo diretório de trabalho da sessão ativa:
popBR2012 <- ler_popbr("POPBR12.DBF")
head(popBR2012)
xtabs(populacao ~ fxetar5, data = popBR2012)

# Um arquivo no diretório FTP do DATASUS
popBR2012 <- ler_popbr("popBR12", source = "url")
xtabs(populacao ~ fxetar5, data = popBR2012)
}

}
