% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomesgruposCSAP.R
\name{nomesgruposCSAP}
\alias{nomesgruposCSAP}
\title{Nomes dos grupos de causa da Lista Brasileira de Condições Sensíveis à Atenção Primária}
\usage{
nomesgruposCSAP(
  lista = "MS",
  lang = "pt.ca",
  classe = "vetor",
  numgrupo = FALSE
)
}
\arguments{
\item{lista}{Lista de causas a ser considerada; pode ser \code{"MS"} (default) para a lista publicada em portaria pelo Ministério da Saúde do Brasil ou "Alfradique" para a lista publicada no artigo de Alfradique et al.}

\item{lang}{idioma em que se apresentam os nomes dos grupos; pode ser: "pt.ca" (default) para nomes em português com acentos; "pt.sa" para nomes em português sem acentos; "en" para nomes em inglês; ou "es" para nomes em castelhano.}

\item{classe}{O output da função deve ser (1) um vetor com a lista dos nomes (padrão, definido por \code{"vetor"}, \code{"v"} ou \code{1}) ou (2) um "data frame" com uma variável com o código do grupo ("g01", etc.) e outra com o nome  definido por \code{"data.frame"}, \code{"df"} ou \code{2})?}

\item{numgrupo}{No caso de se definir um "data frame" no parâmetro \code{classe}, a variável com o nome do grupo deve iniciar com o número do grupo? (v. exemplos).}
}
\value{
Um vetor da classe \code{character} ou uma tabela na classe \code{data frame} com os nomes (abreviados) dos grupos de causa segundo a lista definida pelo usuário.
}
\description{
Lista os grupos de causa da Lista Brasileira de Condições Sensíveis à Atenção Primária, segundo a Portaria do Ministério da Saúde do Brasil, com 19 grupos, ou segundo a publicação em Alfradique et al. (2009), com 20 grupos. Facilita a inclusão ddo grupo de causa como uma variável de um banco de dados. O texto pode ser apresentado em português, espanhol ou inglês.
}
\examples{
nomesgruposCSAP()
nomesgruposCSAP(classe = "df")
nomesgruposCSAP(classe = "df", numgrupo = TRUE)
nomesgruposCSAP(lang = "pt.sa")
nomesgruposCSAP(lang = "en")
nomesgruposCSAP(lang = "es")
nomesgruposCSAP(lista = 'Alfradique')
nomesgruposCSAP(lista = 'Alfradique', classe = "df")
nomesgruposCSAP(lista = 'Alfradique', lang = 'es',
                classe = 'df', numgrupo = TRUE)

# Uso de `classe = 'df'`
require(dplyr)
## Inclui o nome do grupo como uma variável no banco de dados:
aih100 \%>\%
  csapAIH() \%>\%
  filter(csap == "sim") \%>\%
  select(c(4, 6, 9:11)) \%>\%
  left_join(nomesgruposCSAP(classe = 'df'))

left_join(
  csapAIH(aih500, lista = "Alfradique"),
  nomesgruposCSAP(classe = 2, lista = "Alfradique", lang = "en", numgrupo = TRUE)) \%>\%
  group_by(csap, grupo, nomegrupo) \%>\%
  reframe(n()) \%>\%
  print(n = 21)

}
\references{
Alfradique ME et al. Internações por condições sensíveis à atenção primária: a construção da lista brasileira como ferramenta para medir o desempenho do sistema de saúde (Projeto ICSAP - Brasil). Cadernos de Saúde Pública. 2009; 25(6):1337-1349. https://doi.org/10.1590/S0102-311X2009000600016.

Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}
}
\seealso{
\code{\link{csapAIH}}, \code{\link{descreveCSAP}}, \code{\link{desenhaCSAP}}
}
