% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomesgruposCSAP.R
\name{nomesgruposCSAP}
\alias{nomesgruposCSAP}
\title{Nomes dos grupos de causa da Lista Brasileira de Condicoes Sensiveis a Atencao Primaria}
\usage{
nomesgruposCSAP()
}
\value{
Um vetor da classe \code{character} com os nomes (abreviados) dos 19 grupos de causa segundo a Lista Brasileira.
}
\description{
Lista os 19 grupos de causa CSAP, em ordem crescente.
}
\examples{
nomesgruposCSAP()

}
\seealso{
\code{\link{csapAIH}}, \code{\link{descreveCSAP}}, \code{\link{desenhaCSAP}}
}
