% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_popbr.R
\name{popbr}
\alias{popbr}
\title{Estimativas populacionais para os municípios brasileiros.}
\usage{
popbr(x, uf = NULL, municipio = NULL, idade = FALSE)
}
\arguments{
\item{x}{Ano ou vetor com os anos a serem lidos.}

\item{uf}{Unidade(s) da Federação de interesse para seleção. O padrão é todas.}

\item{municipio}{Município(s) de interesse para seleção. O padrão é todos.}

\item{idade}{Argumento lógico. Se TRUE, a idade detalhada é incluída como uma das variáveis. O padrão é FALSE.}
}
\description{
Lê os arquivos com estimativas e contagens da população dos municípios brasileiros por sexo e faixa etária disponibilizados pelo DATASUS.
}
\examples{
# Arquivos no diretório FTP do DATASUS
popbr(2024) |> head()
popbr(2024, idade = TRUE) |> head()
anos <- popbr(2017:2019)
xtabs(populacao ~ fxetar5 + sexo + ano, anos) |> ftable(col.vars = c("ano", "sexo"))
popbr(c(2017, 2019))  |> str()
popbr(2022, "RS") |> head()
popsul22 <- popbr(2022, c("PR", "SC", "RS"))
xtabs(populacao ~ fxetar5 + sexo + UF_SIGLA, popsul22) |> ftable(col.vars = c("UF_SIGLA", "sexo"))
popbr(2013, municipio = "430520") |> head()
popcap <- popbr(2013, municipio = c("431490", "420540"))
xtabs(populacao ~ fxetar5 + sexo + munic_res, popcap) |> ftable(col.vars = c("munic_res", "sexo"))

# Até 2012 a estrutura era outra
popbr(c(1980, 2012))  |> str()
# Por isso o exemplo seguinte dá erro (e ainda não foi trabalhado na função):
\dontrun{
popbr(2012:2013)
}

}
