% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_popbr.R
\name{popbr}
\alias{popbr}
\title{Estimativas populacionais para os municípios brasileiros.}
\usage{
popbr(x, uf = NULL, municipio = NULL, idade = FALSE)
}
\arguments{
\item{x}{Ano ou vetor com os anos a serem lidos.}

\item{uf}{Unidade(s) da Federação de interesse para seleção. O padrão é todas.}

\item{municipio}{Município(s) de interesse para seleção. O padrão é todos. Ainda não implementado.}

\item{idade}{Argumento lógico. A idade detalhada deve ser apresentada? O padrão é FALSE (v. Details).}
}
\description{
Lê os arquivos com estimativas e contagens da população dos municípios brasileiros por sexo e faixa etária disponibilizados pelo DATASUS.
}
\examples{
popbr(2017:2019) |> str()
popbr(c(2017, 2019))  |> str()
popbr(2022, "SC") |> head()
popbr(2010, "RS", idade = TRUE) |> head()
# O exemplo seguinte dá erro, porque a estrutura da base muda em 2013.
\dontrun{
popbr(2012:2013)
}

}
