% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabCSAP.R
\name{tabCSAP}
\alias{tabCSAP}
\title{Tabulate Primary Health Care Sensitive Conditions (PHCSC)}
\usage{
tabCSAP(x, digits = 2, lang = "pt.ca", format = FALSE)
}
\arguments{
\item{x}{Um vetor da classe \code{factor} com os grupos de causa CSAP, nomeados de acordo com o resultado da função \code{\link{csapAIH}}. Esse vetor não precisa ser gerado pela função \code{\link{csapAIH}}, mas deve conter todos os 19 grupos de causa, ainda que sua frequência seja zero, e também devem ser rotulados da mesma forma e ordem que na função, isto é, "g01", "g02", ..., "g19".}

\item{digits}{número de decimais para o arredondamento (com \code{\link{round}}).}

\item{lang}{idioma em que se apresentam os nomes dos grupos; pode ser: "pt.ca" (default) para nomes em português com acentos; "pt.sa" para nomes em português sem acentos; "en" para nomes em inglês; ou "es" para nomes em castelhano.}

\item{format}{Should the table be formatted to print? (default = FALSE).}
}
\value{
Uma tabela com a frequência absoluta dos grupos de causa e sua distribuição proporcional sobre o total de internações e sobre o total de ICSAP. Se os grupos forem classificados segundo a Lista Brasileira publicada em Portaria Ministerial, a tabela terá
}
\description{
Tabulates Primary Health Care Sensitive Conditions (PHCSC, CSAP) according to the Brazilian list of causes.
}
\examples{
data("aih500")
tabCSAP(csapAIH(aih500)$grupo)
tabCSAP(csapAIH(aih500)$grupo, lang = "pt.sa")
tabCSAP(csapAIH(aih500)$grupo, lang = "en")
tabCSAP(csapAIH(aih500)$grupo, lang = "es")
tabCSAP(csapAIH(aih500, lista = "Alfradique")$grupo)
tabCSAP(csapAIH(aih500, lista = "Alfradique")$grupo,lang = "pt.sa")
tabCSAP(csapAIH(aih500, lista = "Alfradique")$grupo,lang = "en")
tabCSAP(csapAIH(aih500, lista = "Alfradique")$grupo,lang = "es")

}
\references{
Brasil. Ministério da Saúde. Secretaria de Atenção à Saúde. Portaria No 221, de 17 de abril de 2008. \url{http://bvsms.saude.gov.br/bvs/saudelegis/sas/2008/prt0221_17_04_2008.html}
}
\seealso{
\code{\link{csapAIH}}, \code{\link{nomesgruposCSAP}}
}
