% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufbr.R
\name{ufbr}
\alias{ufbr}
\title{Lista as Unidades da Federação por Região do Brasil}
\usage{
ufbr(nomes = FALSE)
}
\arguments{
\item{nomes}{Se \code{nomes = TRUE}, é incluída uma coluna com os nomes dos Estados. Se \code{nomes = FALSE} (padrão), a UF é identificada apenas pela sigla.}
}
\value{
Um objeto da classe \code{data.frame}
}
\description{
Lista as siglas e nomes das Unidades Federativas (UF) por Região do Brasil segundo seu código no IBGE, permitindo a decodificação do código do IBGE para a sigla ou nome da Unidade da Federação.
}
\examples{
ufbr()
ufbr(nomes = TRUE)

}
